Crate vkvk

Modules

handles

non_dispatchable_handles

prelude

vk_platform

vk_result

vk_structure_type

Structs

VkAccessFlagBits
VkAccessFlagBits

VkBaseInStructure
VkBaseInStructure

VkBaseOutStructure
VkBaseOutStructure

VkBool32
The “truthy” type for Vulkan.

VkBufferMemoryBarrier
VkBufferMemoryBarrier

VkDeviceAddress
VkDeviceAddress

VkDeviceSize
VkDeviceSize

VkDispatchIndirectCommand
VkDispatchIndirectCommand

VkDrawIndexedIndirectCommand
VkDrawIndexedIndirectCommand

VkDrawIndirectCommand
VkDrawIndirectCommand

VkExtent2D
VkExtent2D

VkExtent3D
VkExtent3D

VkImageAspectFlagBits
VkImageAspectFlagBits

VkImageLayout
VkImageLayout

VkImageMemoryBarrier
VkImageMemoryBarrier

VkImageSubresourceRange
VkImageSubresourceRange

VkMemoryBarrier
VkMemoryBarrier

VkObjectType
Enums to track objects of various types - also see objtypeenum attributes on type tags.

VkOffset2D
VkOffset2D

VkOffset3D
VkOffset3D

VkPipelineCacheHeaderVersion
VkPipelineCacheHeaderVersion

VkRect2D
VkRect2D

VkVendorId
VkVendorId

VulkanVersion
Provides simple access to a vulkan version value.

Constants

VK_ACCESS_ACCELERATION_STRUCTURE_READ_BIT_KHR
Provided by VK_KHR_acceleration_structure

VK_ACCESS_ACCELERATION_STRUCTURE_READ_BIT_NV
Provided by VK_NV_ray_tracing

VK_ACCESS_ACCELERATION_STRUCTURE_WRITE_BIT_KHR
Provided by VK_KHR_acceleration_structure

VK_ACCESS_ACCELERATION_STRUCTURE_WRITE_BIT_NV
Provided by VK_NV_ray_tracing

VK_ACCESS_COLOR_ATTACHMENT_READ_BIT
Controls coherency of color attachment reads

VK_ACCESS_COLOR_ATTACHMENT_READ_NONCOHERENT_BIT_EXT
Provided by VK_EXT_blend_operation_advanced

VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT
Controls coherency of color attachment writes

VK_ACCESS_COMMAND_PREPROCESS_READ_BIT_NV
Provided by VK_NV_device_generated_commands

VK_ACCESS_COMMAND_PREPROCESS_WRITE_BIT_NV
Provided by VK_NV_device_generated_commands

VK_ACCESS_CONDITIONAL_RENDERING_READ_BIT_EXT
Provided by VK_EXT_conditional_rendering

VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_READ_BIT
Controls coherency of depth/stencil attachment reads

VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT
Controls coherency of depth/stencil attachment writes

VK_ACCESS_FRAGMENT_DENSITY_MAP_READ_BIT_EXT
Provided by VK_EXT_fragment_density_map

VK_ACCESS_FRAGMENT_SHADING_RATE_ATTACHMENT_READ_BIT_KHR
Provided by VK_KHR_fragment_shading_rate

VK_ACCESS_HOST_READ_BIT
Controls coherency of host reads

VK_ACCESS_HOST_WRITE_BIT
Controls coherency of host writes

VK_ACCESS_INDEX_READ_BIT
Controls coherency of index reads

VK_ACCESS_INDIRECT_COMMAND_READ_BIT
Controls coherency of indirect command reads

VK_ACCESS_INPUT_ATTACHMENT_READ_BIT
Controls coherency of input attachment reads

VK_ACCESS_MEMORY_READ_BIT
Controls coherency of memory reads

VK_ACCESS_MEMORY_WRITE_BIT
Controls coherency of memory writes

VK_ACCESS_NONE_KHR
Provided by VK_KHR_synchronization2

VK_ACCESS_SHADER_READ_BIT
Controls coherency of shader reads

VK_ACCESS_SHADER_WRITE_BIT
Controls coherency of shader writes

VK_ACCESS_SHADING_RATE_IMAGE_READ_BIT_NV
Provided by VK_NV_shading_rate_image

VK_ACCESS_TRANSFER_READ_BIT
Controls coherency of transfer reads

VK_ACCESS_TRANSFER_WRITE_BIT
Controls coherency of transfer writes

VK_ACCESS_TRANSFORM_FEEDBACK_COUNTER_READ_BIT_EXT
Provided by VK_EXT_transform_feedback

VK_ACCESS_TRANSFORM_FEEDBACK_COUNTER_WRITE_BIT_EXT
Provided by VK_EXT_transform_feedback

VK_ACCESS_TRANSFORM_FEEDBACK_WRITE_BIT_EXT
Provided by VK_EXT_transform_feedback

VK_ACCESS_UNIFORM_READ_BIT
Controls coherency of uniform buffer reads

VK_ACCESS_VERTEX_ATTRIBUTE_READ_BIT
Controls coherency of vertex attribute reads

VK_ATTACHMENT_UNUSED

VK_FALSE

VK_IMAGE_ASPECT_COLOR_BIT

VK_IMAGE_ASPECT_DEPTH_BIT

VK_IMAGE_ASPECT_MEMORY_PLANE_0_BIT_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_IMAGE_ASPECT_MEMORY_PLANE_1_BIT_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_IMAGE_ASPECT_MEMORY_PLANE_2_BIT_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_IMAGE_ASPECT_MEMORY_PLANE_3_BIT_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_IMAGE_ASPECT_METADATA_BIT

VK_IMAGE_ASPECT_PLANE_0_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_ASPECT_PLANE_0_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_IMAGE_ASPECT_PLANE_1_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_ASPECT_PLANE_1_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_IMAGE_ASPECT_PLANE_2_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_ASPECT_PLANE_2_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_IMAGE_ASPECT_STENCIL_BIT

VK_IMAGE_LAYOUT_ATTACHMENT_OPTIMAL_KHR
Provided by VK_KHR_synchronization2

VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL
Optimal layout when image is only used for color attachment read/write

VK_IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL
Provided by VK_VERSION_1_2

VK_IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL_KHR
Provided by VK_KHR_separate_depth_stencil_layouts

VK_IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL
Provided by VK_VERSION_1_1

VK_IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL_KHR
Provided by VK_KHR_maintenance2

VK_IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL
Provided by VK_VERSION_1_2

VK_IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL_KHR
Provided by VK_KHR_separate_depth_stencil_layouts

VK_IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL
Provided by VK_VERSION_1_1

VK_IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL_KHR
Provided by VK_KHR_maintenance2

VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL
Optimal layout when image is only used for depth/stencil attachment read/write

VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL
Optimal layout when image is used for read only depth/stencil attachment and shader access

VK_IMAGE_LAYOUT_FRAGMENT_DENSITY_MAP_OPTIMAL_EXT
Provided by VK_EXT_fragment_density_map

VK_IMAGE_LAYOUT_FRAGMENT_SHADING_RATE_ATTACHMENT_OPTIMAL_KHR
Provided by VK_KHR_fragment_shading_rate

VK_IMAGE_LAYOUT_GENERAL
General layout when image can be used for any kind of access

VK_IMAGE_LAYOUT_PREINITIALIZED
Initial layout used when the data is populated by the CPU

VK_IMAGE_LAYOUT_PRESENT_SRC_KHR
Provided by VK_KHR_swapchain

VK_IMAGE_LAYOUT_READ_ONLY_OPTIMAL_KHR
Provided by VK_KHR_synchronization2

VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL
Optimal layout when image is used for read only shader access

VK_IMAGE_LAYOUT_SHADING_RATE_OPTIMAL_NV
Provided by VK_NV_shading_rate_image

VK_IMAGE_LAYOUT_SHARED_PRESENT_KHR
Provided by VK_KHR_shared_presentable_image

VK_IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL
Provided by VK_VERSION_1_2

VK_IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL_KHR
Provided by VK_KHR_separate_depth_stencil_layouts

VK_IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL
Provided by VK_VERSION_1_2

VK_IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL_KHR
Provided by VK_KHR_separate_depth_stencil_layouts

VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL
Optimal layout when image is used only as destination of transfer operations

VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL
Optimal layout when image is used only as source of transfer operations

VK_IMAGE_LAYOUT_UNDEFINED
Implicit layout an image is when its contents are undefined due to various reasons (e.g. right after creation)

VK_LOD_CLAMP_NONE

VK_OBJECT_TYPE_ACCELERATION_STRUCTURE_KHR
Provided by VK_KHR_acceleration_structure

VK_OBJECT_TYPE_ACCELERATION_STRUCTURE_NV
Provided by VK_NV_ray_tracing

VK_OBJECT_TYPE_BUFFER

VK_OBJECT_TYPE_BUFFER_VIEW

VK_OBJECT_TYPE_COMMAND_BUFFER

VK_OBJECT_TYPE_COMMAND_POOL

VK_OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT
Provided by VK_EXT_debug_report

VK_OBJECT_TYPE_DEBUG_UTILS_MESSENGER_EXT
Provided by VK_EXT_debug_utils

VK_OBJECT_TYPE_DEFERRED_OPERATION_KHR
Provided by VK_KHR_deferred_host_operations

VK_OBJECT_TYPE_DESCRIPTOR_POOL

VK_OBJECT_TYPE_DESCRIPTOR_SET

VK_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT

VK_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE
Provided by VK_VERSION_1_1

VK_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_KHR
Provided by VK_KHR_descriptor_update_template

VK_OBJECT_TYPE_DEVICE

VK_OBJECT_TYPE_DEVICE_MEMORY

VK_OBJECT_TYPE_DISPLAY_KHR
Provided by VK_KHR_display

VK_OBJECT_TYPE_DISPLAY_MODE_KHR
Provided by VK_KHR_display

VK_OBJECT_TYPE_EVENT

VK_OBJECT_TYPE_FENCE

VK_OBJECT_TYPE_FRAMEBUFFER

VK_OBJECT_TYPE_IMAGE

VK_OBJECT_TYPE_IMAGE_VIEW

VK_OBJECT_TYPE_INDIRECT_COMMANDS_LAYOUT_NV
Provided by VK_NV_device_generated_commands

VK_OBJECT_TYPE_INSTANCE

VK_OBJECT_TYPE_PERFORMANCE_CONFIGURATION_INTEL
Provided by VK_INTEL_performance_query

VK_OBJECT_TYPE_PHYSICAL_DEVICE

VK_OBJECT_TYPE_PIPELINE

VK_OBJECT_TYPE_PIPELINE_CACHE

VK_OBJECT_TYPE_PIPELINE_LAYOUT

VK_OBJECT_TYPE_PRIVATE_DATA_SLOT_EXT
Provided by VK_EXT_private_data

VK_OBJECT_TYPE_QUERY_POOL

VK_OBJECT_TYPE_QUEUE

VK_OBJECT_TYPE_RENDER_PASS

VK_OBJECT_TYPE_SAMPLER

VK_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION
Provided by VK_VERSION_1_1

VK_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_OBJECT_TYPE_SEMAPHORE

VK_OBJECT_TYPE_SHADER_MODULE

VK_OBJECT_TYPE_SURFACE_KHR
Provided by VK_KHR_surface

VK_OBJECT_TYPE_SWAPCHAIN_KHR
Provided by VK_KHR_swapchain

VK_OBJECT_TYPE_UNKNOWN

VK_OBJECT_TYPE_VALIDATION_CACHE_EXT
Provided by VK_EXT_validation_cache

VK_PIPELINE_CACHE_HEADER_VERSION_ONE

VK_QUEUE_FAMILY_IGNORED

VK_REMAINING_ARRAY_LAYERS

VK_REMAINING_MIP_LEVELS

VK_SUBPASS_EXTERNAL

VK_TRUE

VK_VENDOR_ID_CODEPLAY
Codeplay Software Ltd. vendor ID

VK_VENDOR_ID_KAZAN
Kazan Software Renderer

VK_VENDOR_ID_MESA
Mesa vendor ID

VK_VENDOR_ID_POCL
PoCL vendor ID

VK_VENDOR_ID_VIV
Vivante vendor ID

VK_VENDOR_ID_VSI
VeriSilicon vendor ID

VK_WHOLE_SIZE

Type Definitions

VkAccessFlags

VkImageAspectFlags