Module vkvk::flag_bits

Vulkan types where one or more flags are bit-packed together.

These types support the standard bitwise operations for enabling and disabling specific bits:

Add a new flag to a base value: base | new

Remove a flag from a base value: base & (!removal)

Structs

VkAccessFlagBits
VkAccessFlagBits

VkAttachmentDescriptionFlagBits
VkAttachmentDescriptionFlagBits

VkBufferCreateFlagBits
VkBufferCreateFlagBits

VkBufferUsageFlagBits
VkBufferUsageFlagBits

VkBufferViewCreateFlagBits
VkBufferViewCreateFlags

VkColorComponentFlagBits
VkColorComponentFlagBits

VkCommandBufferResetFlagBits
VkCommandBufferResetFlagBits

VkCommandBufferUsageFlagBits
VkCommandBufferUsageFlagBits

VkCommandPoolCreateFlagBits
VkCommandPoolCreateFlagBits

VkCommandPoolResetFlagBits
VkCommandPoolResetFlagBits

VkCullModeFlagBits
VkCullModeFlagBits

VkDependencyFlagBits
VkDependencyFlagBits

VkDescriptorPoolCreateFlagBits
VkDescriptorPoolCreateFlagBits

VkDescriptorPoolResetFlagBits
VkXlibSurfaceCreateFlagsKHR

VkDescriptorSetLayoutCreateFlagBits
VkDescriptorSetLayoutCreateFlagBits

VkDeviceCreateFlagBits
VkDeviceCreateFlags

VkDeviceQueueCreateFlagBits
VkDeviceQueueCreateFlagBits

VkEventCreateFlagBits
VkEventCreateFlags

VkFenceCreateFlagBits
VkFenceCreateFlagBits

VkFormatFeatureFlagBits
VkFormatFeatureFlagBits

VkFramebufferCreateFlagBits
VkFramebufferCreateFlagBits

VkImageAspectFlagBits
VkImageAspectFlagBits

VkImageCreateFlagBits
VkImageCreateFlagBits

VkImageUsageFlagBits
VkImageUsageFlagBits

VkImageViewCreateFlagBits
VkImageViewCreateFlagBits

VkInstanceCreateFlagBits
VkInstanceCreateFlags

VkMemoryHeapFlagBits
VkMemoryHeapFlagBits

VkMemoryMapFlagBits
VkXlibSurfaceCreateFlagsKHR

VkMemoryPropertyFlagBits
VkMemoryPropertyFlagBits

VkPipelineCacheCreateFlagBits
VkPipelineCacheCreateFlagBits

VkPipelineColorBlendStateCreateFlagBits
VkPipelineColorBlendStateCreateFlags

VkPipelineCreateFlagBits
VkPipelineCreateFlagBits

VkPipelineDepthStencilStateCreateFlagBits
VkPipelineDepthStencilStateCreateFlags

VkPipelineDynamicStateCreateFlagBits
VkPipelineDynamicStateCreateFlags

VkPipelineInputAssemblyStateCreateFlagBits
VkPipelineInputAssemblyStateCreateFlags

VkPipelineLayoutCreateFlagBits
VkPipelineLayoutCreateFlags

VkPipelineMultisampleStateCreateFlagBits
VkPipelineMultisampleStateCreateFlags

VkPipelineRasterizationStateCreateFlagBits
VkPipelineRasterizationStateCreateFlags

VkPipelineShaderStageCreateFlagBits
VkPipelineShaderStageCreateFlagBits

VkPipelineStageFlagBits
VkPipelineStageFlagBits

VkPipelineTessellationStateCreateFlagBits
VkPipelineTessellationStateCreateFlags

VkPipelineVertexInputStateCreateFlagBits
VkPipelineVertexInputStateCreateFlags

VkPipelineViewportStateCreateFlagBits
VkPipelineViewportStateCreateFlags

VkQueryControlFlagBits
VkQueryControlFlagBits

VkQueryPipelineStatisticFlagBits
VkQueryPipelineStatisticFlagBits

VkQueryPoolCreateFlagBits
VkQueryPoolCreateFlags

VkQueryResultFlagBits
VkQueryResultFlagBits

VkQueueFlagBits
VkQueueFlagBits

VkRenderPassCreateFlagBits
VkRenderPassCreateFlagBits

VkSampleCountFlagBits
VkSampleCountFlagBits

VkSamplerCreateFlagBits
VkSamplerCreateFlagBits

VkSemaphoreCreateFlagBits
VkSemaphoreCreateFlags

VkShaderModuleCreateFlagBits
VkShaderModuleCreateFlagBits

VkShaderStageFlagBits
VkShaderStageFlagBits

VkSparseImageFormatFlagBits
VkSparseImageFormatFlagBits

VkSparseMemoryBindFlagBits
VkSparseMemoryBindFlagBits

VkStencilFaceFlagBits
VkStencilFaceFlagBits

VkSubpassDescriptionFlagBits
VkSubpassDescriptionFlagBits

Constants

VK_ACCESS_ACCELERATION_STRUCTURE_READ_BIT_KHR

VK_ACCESS_ACCELERATION_STRUCTURE_READ_BIT_NV

VK_ACCESS_ACCELERATION_STRUCTURE_WRITE_BIT_KHR

VK_ACCESS_ACCELERATION_STRUCTURE_WRITE_BIT_NV

VK_ACCESS_COLOR_ATTACHMENT_READ_BIT
Controls coherency of color attachment reads

VK_ACCESS_COLOR_ATTACHMENT_READ_NONCOHERENT_BIT_EXT

VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT
Controls coherency of color attachment writes

VK_ACCESS_COMMAND_PREPROCESS_READ_BIT_NV

VK_ACCESS_COMMAND_PREPROCESS_WRITE_BIT_NV

VK_ACCESS_CONDITIONAL_RENDERING_READ_BIT_EXT

VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_READ_BIT
Controls coherency of depth/stencil attachment reads

VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT
Controls coherency of depth/stencil attachment writes

VK_ACCESS_FRAGMENT_DENSITY_MAP_READ_BIT_EXT

VK_ACCESS_FRAGMENT_SHADING_RATE_ATTACHMENT_READ_BIT_KHR

VK_ACCESS_HOST_READ_BIT
Controls coherency of host reads

VK_ACCESS_HOST_WRITE_BIT
Controls coherency of host writes

VK_ACCESS_INDEX_READ_BIT
Controls coherency of index reads

VK_ACCESS_INDIRECT_COMMAND_READ_BIT
Controls coherency of indirect command reads

VK_ACCESS_INPUT_ATTACHMENT_READ_BIT
Controls coherency of input attachment reads

VK_ACCESS_MEMORY_READ_BIT
Controls coherency of memory reads

VK_ACCESS_MEMORY_WRITE_BIT
Controls coherency of memory writes

VK_ACCESS_NONE_KHR

VK_ACCESS_SHADER_READ_BIT
Controls coherency of shader reads

VK_ACCESS_SHADER_WRITE_BIT
Controls coherency of shader writes

VK_ACCESS_SHADING_RATE_IMAGE_READ_BIT_NV

VK_ACCESS_TRANSFER_READ_BIT
Controls coherency of transfer reads

VK_ACCESS_TRANSFER_WRITE_BIT
Controls coherency of transfer writes

VK_ACCESS_TRANSFORM_FEEDBACK_COUNTER_READ_BIT_EXT

VK_ACCESS_TRANSFORM_FEEDBACK_COUNTER_WRITE_BIT_EXT

VK_ACCESS_TRANSFORM_FEEDBACK_WRITE_BIT_EXT

VK_ACCESS_UNIFORM_READ_BIT
Controls coherency of uniform buffer reads

VK_ACCESS_VERTEX_ATTRIBUTE_READ_BIT
Controls coherency of vertex attribute reads

VK_ATTACHMENT_DESCRIPTION_MAY_ALIAS_BIT
The attachment may alias physical memory of another attachment in the same render pass

VK_BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT
Provided by VK_VERSION_1_2

VK_BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT_EXT
Provided by VK_EXT_buffer_device_address

VK_BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT_KHR
Provided by VK_KHR_buffer_device_address

VK_BUFFER_CREATE_PROTECTED_BIT
Provided by VK_VERSION_1_1

VK_BUFFER_CREATE_SPARSE_ALIASED_BIT
Buffer should support constant data access to physical memory ranges mapped into multiple locations of sparse buffers

VK_BUFFER_CREATE_SPARSE_BINDING_BIT
Buffer should support sparse backing

VK_BUFFER_CREATE_SPARSE_RESIDENCY_BIT
Buffer should support sparse backing with partial residency

VK_BUFFER_USAGE_ACCELERATION_STRUCTURE_BUILD_INPUT_READ_ONLY_BIT_KHR
Provided by VK_KHR_acceleration_structure

VK_BUFFER_USAGE_ACCELERATION_STRUCTURE_STORAGE_BIT_KHR
Provided by VK_KHR_acceleration_structure

VK_BUFFER_USAGE_CONDITIONAL_RENDERING_BIT_EXT
Provided by VK_EXT_conditional_rendering

VK_BUFFER_USAGE_INDEX_BUFFER_BIT
Can be used as source of fixed-function index fetch (index buffer)

VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT
Can be the source of indirect parameters (e.g. indirect buffer, parameter buffer)

VK_BUFFER_USAGE_RAY_TRACING_BIT_NV
Provided by VK_NV_ray_tracing

VK_BUFFER_USAGE_SHADER_BINDING_TABLE_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT
Provided by VK_VERSION_1_2

VK_BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT_EXT
Provided by VK_EXT_buffer_device_address

VK_BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT_KHR
Provided by VK_KHR_buffer_device_address

VK_BUFFER_USAGE_STORAGE_BUFFER_BIT
Can be used as SSBO

VK_BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT
Can be used as IBO

VK_BUFFER_USAGE_TRANSFER_DST_BIT
Can be used as a destination of transfer operations

VK_BUFFER_USAGE_TRANSFER_SRC_BIT
Can be used as a source of transfer operations

VK_BUFFER_USAGE_TRANSFORM_FEEDBACK_BUFFER_BIT_EXT
Provided by VK_EXT_transform_feedback

VK_BUFFER_USAGE_TRANSFORM_FEEDBACK_COUNTER_BUFFER_BIT_EXT
Provided by VK_EXT_transform_feedback

VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT
Can be used as UBO

VK_BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT
Can be used as TBO

VK_BUFFER_USAGE_VERTEX_BUFFER_BIT
Can be used as source of fixed-function vertex fetch (VBO)

VK_COLOR_COMPONENT_A_BIT

VK_COLOR_COMPONENT_B_BIT

VK_COLOR_COMPONENT_G_BIT

VK_COLOR_COMPONENT_R_BIT

VK_COMMAND_BUFFER_RESET_RELEASE_RESOURCES_BIT
Release resources owned by the buffer

VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT

VK_COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT

VK_COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT
Command buffer may be submitted/executed more than once simultaneously

VK_COMMAND_POOL_CREATE_PROTECTED_BIT
Provided by VK_VERSION_1_1

VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT
Command buffers may release their memory individually

VK_COMMAND_POOL_CREATE_TRANSIENT_BIT
Command buffers have a short lifetime

VK_COMMAND_POOL_RESET_RELEASE_RESOURCES_BIT
Release resources owned by the pool

VK_CULL_MODE_BACK_BIT

VK_CULL_MODE_FRONT_AND_BACK

VK_CULL_MODE_FRONT_BIT

VK_CULL_MODE_NONE

VK_DEPENDENCY_BY_REGION_BIT
Dependency is per pixel region

VK_DEPENDENCY_DEVICE_GROUP_BIT
Provided by VK_VERSION_1_1

VK_DEPENDENCY_DEVICE_GROUP_BIT_KHR
Provided by VK_KHR_device_group

VK_DEPENDENCY_VIEW_LOCAL_BIT
Provided by VK_VERSION_1_1

VK_DEPENDENCY_VIEW_LOCAL_BIT_KHR
Provided by VK_KHR_multiview

VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT
Descriptor sets may be freed individually

VK_DESCRIPTOR_POOL_CREATE_HOST_ONLY_BIT_VALVE
Provided by VK_VALVE_mutable_descriptor_type

VK_DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT
Provided by VK_VERSION_1_2

VK_DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT_EXT

VK_DESCRIPTOR_SET_LAYOUT_CREATE_HOST_ONLY_POOL_BIT_VALVE
Provided by VK_VALVE_mutable_descriptor_type

VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR
Provided by VK_KHR_push_descriptor

VK_DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT
Provided by VK_VERSION_1_2

VK_DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT_EXT
Provided by VK_EXT_descriptor_indexing

VK_FENCE_CREATE_SIGNALED_BIT

VK_FORMAT_FEATURE_ACCELERATION_STRUCTURE_VERTEX_BUFFER_BIT_KHR
Provided by VK_KHR_acceleration_structure

VK_FORMAT_FEATURE_BLIT_DST_BIT
Format can be used as the destination image of blits with vkCmdBlitImage

VK_FORMAT_FEATURE_BLIT_SRC_BIT
Format can be used as the source image of blits with vkCmdBlitImage

VK_FORMAT_FEATURE_COLOR_ATTACHMENT_BIT
Format can be used for color attachment images

VK_FORMAT_FEATURE_COLOR_ATTACHMENT_BLEND_BIT
Format supports blending in case it is used for color attachment images

VK_FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT
Format can be used for depth/stencil attachment images

VK_FORMAT_FEATURE_DISJOINT_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_DISJOINT_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_FRAGMENT_DENSITY_MAP_BIT_EXT
Provided by VK_EXT_fragment_density_map

VK_FORMAT_FEATURE_FRAGMENT_SHADING_RATE_ATTACHMENT_BIT_KHR
Provided by VK_KHR_fragment_shading_rate

VK_FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_SAMPLED_IMAGE_BIT
Format can be used for sampled images (SAMPLED_IMAGE and COMBINED_IMAGE_SAMPLER descriptor types)

VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_CUBIC_BIT_EXT
Provided by VK_EXT_filter_cubic

VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_CUBIC_BIT_IMG
Provided by VK_IMG_filter_cubic

VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT
Format can be filtered with VK_FILTER_LINEAR when being sampled

VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_MINMAX_BIT
Provided by VK_VERSION_1_2

VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_MINMAX_BIT_EXT
Provided by VK_EXT_sampler_filter_minmax

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_SEPARATE_RECONSTRUCTION_FILTER_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_SEPARATE_RECONSTRUCTION_FILTER_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_STORAGE_IMAGE_ATOMIC_BIT
Format supports atomic operations in case it is used for storage images

VK_FORMAT_FEATURE_STORAGE_IMAGE_BIT
Format can be used for storage images (STORAGE_IMAGE descriptor type)

VK_FORMAT_FEATURE_STORAGE_TEXEL_BUFFER_ATOMIC_BIT
Format supports atomic operations in case it is used for storage texel buffers

VK_FORMAT_FEATURE_STORAGE_TEXEL_BUFFER_BIT
Format can be used for storage texel buffers (IBOs)

VK_FORMAT_FEATURE_TRANSFER_DST_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_TRANSFER_DST_BIT_KHR
Provided by VK_KHR_maintenance1

VK_FORMAT_FEATURE_TRANSFER_SRC_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_TRANSFER_SRC_BIT_KHR
Provided by VK_KHR_maintenance1

VK_FORMAT_FEATURE_UNIFORM_TEXEL_BUFFER_BIT
Format can be used for uniform texel buffers (TBOs)

VK_FORMAT_FEATURE_VERTEX_BUFFER_BIT
Format can be used for vertex buffers (VBOs)

VK_FRAMEBUFFER_CREATE_IMAGELESS_BIT
Provided by VK_VERSION_1_2

VK_FRAMEBUFFER_CREATE_IMAGELESS_BIT_KHR
Provided by VK_KHR_imageless_framebuffer

VK_IMAGE_ASPECT_COLOR_BIT

VK_IMAGE_ASPECT_DEPTH_BIT

VK_IMAGE_ASPECT_MEMORY_PLANE_0_BIT_EXT

VK_IMAGE_ASPECT_MEMORY_PLANE_1_BIT_EXT

VK_IMAGE_ASPECT_MEMORY_PLANE_2_BIT_EXT

VK_IMAGE_ASPECT_MEMORY_PLANE_3_BIT_EXT

VK_IMAGE_ASPECT_METADATA_BIT

VK_IMAGE_ASPECT_PLANE_0_BIT

VK_IMAGE_ASPECT_PLANE_0_BIT_KHR

VK_IMAGE_ASPECT_PLANE_1_BIT

VK_IMAGE_ASPECT_PLANE_1_BIT_KHR

VK_IMAGE_ASPECT_PLANE_2_BIT

VK_IMAGE_ASPECT_PLANE_2_BIT_KHR

VK_IMAGE_ASPECT_STENCIL_BIT

VK_IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT_KHR
Provided by VK_KHR_maintenance1

VK_IMAGE_CREATE_ALIAS_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_ALIAS_BIT_KHR
Provided by VK_KHR_bind_memory2

VK_IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT_KHR
Provided by VK_KHR_maintenance2

VK_IMAGE_CREATE_CORNER_SAMPLED_BIT_NV
Provided by VK_NV_corner_sampled_image

VK_IMAGE_CREATE_CUBE_COMPATIBLE_BIT
Allows creating image views with cube type from the created image

VK_IMAGE_CREATE_DISJOINT_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_DISJOINT_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_IMAGE_CREATE_EXTENDED_USAGE_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_EXTENDED_USAGE_BIT_KHR
Provided by VK_KHR_maintenance2

VK_IMAGE_CREATE_MUTABLE_FORMAT_BIT
Allows image views to have different format than the base image

VK_IMAGE_CREATE_PROTECTED_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_SAMPLE_LOCATIONS_COMPATIBLE_DEPTH_BIT_EXT
Provided by VK_EXT_sample_locations

VK_IMAGE_CREATE_SPARSE_ALIASED_BIT
Image should support constant data access to physical memory ranges mapped into multiple locations of sparse images

VK_IMAGE_CREATE_SPARSE_BINDING_BIT
Image should support sparse backing

VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT
Image should support sparse backing with partial residency

VK_IMAGE_CREATE_SPLIT_INSTANCE_BIND_REGIONS_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_SPLIT_INSTANCE_BIND_REGIONS_BIT_KHR
Provided by VK_KHR_device_group with VK_KHR_bind_memory2

VK_IMAGE_CREATE_SUBSAMPLED_BIT_EXT
Provided by VK_EXT_fragment_density_map

VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT
Can be used as framebuffer color attachment

VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT
Can be used as framebuffer depth/stencil attachment

VK_IMAGE_USAGE_FRAGMENT_DENSITY_MAP_BIT_EXT
Provided by VK_EXT_fragment_density_map

VK_IMAGE_USAGE_FRAGMENT_SHADING_RATE_ATTACHMENT_BIT_KHR
Provided by VK_KHR_fragment_shading_rate

VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT
Can be used as framebuffer input attachment

VK_IMAGE_USAGE_SAMPLED_BIT
Can be sampled from (SAMPLED_IMAGE and COMBINED_IMAGE_SAMPLER descriptor types)

VK_IMAGE_USAGE_SHADING_RATE_IMAGE_BIT_NV
Provided by VK_NV_shading_rate_image

VK_IMAGE_USAGE_STORAGE_BIT
Can be used as storage image (STORAGE_IMAGE descriptor type)

VK_IMAGE_USAGE_TRANSFER_DST_BIT
Can be used as a destination of transfer operations

VK_IMAGE_USAGE_TRANSFER_SRC_BIT
Can be used as a source of transfer operations

VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT
Image data not needed outside of rendering

VK_IMAGE_VIEW_CREATE_FRAGMENT_DENSITY_MAP_DEFERRED_BIT_EXT
Provided by VK_EXT_fragment_density_map2

VK_IMAGE_VIEW_CREATE_FRAGMENT_DENSITY_MAP_DYNAMIC_BIT_EXT
Provided by VK_EXT_fragment_density_map

VK_MEMORY_HEAP_DEVICE_LOCAL_BIT
If set, heap represents device memory

VK_MEMORY_HEAP_MULTI_INSTANCE_BIT
Provided by VK_VERSION_1_1

VK_MEMORY_HEAP_MULTI_INSTANCE_BIT_KHR
Provided by VK_KHR_device_group_creation

VK_MEMORY_PROPERTY_DEVICE_COHERENT_BIT_AMD
Provided by VK_AMD_device_coherent_memory

VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT
If otherwise stated, then allocate memory on device

VK_MEMORY_PROPERTY_DEVICE_UNCACHED_BIT_AMD
Provided by VK_AMD_device_coherent_memory

VK_MEMORY_PROPERTY_HOST_CACHED_BIT
Memory will be cached by the host

VK_MEMORY_PROPERTY_HOST_COHERENT_BIT
Memory will have i/o coherency. If not set, application may need to use vkFlushMappedMemoryRanges and vkInvalidateMappedMemoryRanges to flush/invalidate host cache

VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT
Memory is mappable by host

VK_MEMORY_PROPERTY_LAZILY_ALLOCATED_BIT
Memory may be allocated by the driver when it is required

VK_MEMORY_PROPERTY_PROTECTED_BIT
Provided by VK_VERSION_1_1

VK_PIPELINE_CACHE_CREATE_EXTERNALLY_SYNCHRONIZED_BIT_EXT
Provided by VK_EXT_pipeline_creation_cache_control

VK_PIPELINE_CREATE_ALLOW_DERIVATIVES_BIT

VK_PIPELINE_CREATE_CAPTURE_INTERNAL_REPRESENTATIONS_BIT_KHR
Provided by VK_KHR_pipeline_executable_properties

VK_PIPELINE_CREATE_CAPTURE_STATISTICS_BIT_KHR
Provided by VK_KHR_pipeline_executable_properties

VK_PIPELINE_CREATE_DEFER_COMPILE_BIT_NV
Provided by VK_NV_ray_tracing

VK_PIPELINE_CREATE_DERIVATIVE_BIT

VK_PIPELINE_CREATE_DISABLE_OPTIMIZATION_BIT

VK_PIPELINE_CREATE_DISPATCH_BASE

VK_PIPELINE_CREATE_DISPATCH_BASE_BIT
Provided by VK_VERSION_1_1

VK_PIPELINE_CREATE_DISPATCH_BASE_KHR

VK_PIPELINE_CREATE_EARLY_RETURN_ON_FAILURE_BIT_EXT
Provided by VK_EXT_pipeline_creation_cache_control

VK_PIPELINE_CREATE_FAIL_ON_PIPELINE_COMPILE_REQUIRED_BIT_EXT
Provided by VK_EXT_pipeline_creation_cache_control

VK_PIPELINE_CREATE_INDIRECT_BINDABLE_BIT_NV
Provided by VK_NV_device_generated_commands

VK_PIPELINE_CREATE_LIBRARY_BIT_KHR
Provided by VK_KHR_pipeline_library

VK_PIPELINE_CREATE_RAY_TRACING_NO_NULL_ANY_HIT_SHADERS_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_PIPELINE_CREATE_RAY_TRACING_NO_NULL_CLOSEST_HIT_SHADERS_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_PIPELINE_CREATE_RAY_TRACING_NO_NULL_INTERSECTION_SHADERS_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_PIPELINE_CREATE_RAY_TRACING_NO_NULL_MISS_SHADERS_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_PIPELINE_CREATE_RAY_TRACING_SHADER_GROUP_HANDLE_CAPTURE_REPLAY_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_PIPELINE_CREATE_RAY_TRACING_SKIP_AABBS_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_PIPELINE_CREATE_RAY_TRACING_SKIP_TRIANGLES_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_PIPELINE_CREATE_VIEW_INDEX_FROM_DEVICE_INDEX_BIT
Provided by VK_VERSION_1_1

VK_PIPELINE_CREATE_VIEW_INDEX_FROM_DEVICE_INDEX_BIT_KHR

VK_PIPELINE_SHADER_STAGE_CREATE_ALLOW_VARYING_SUBGROUP_SIZE_BIT_EXT
Provided by VK_EXT_subgroup_size_control

VK_PIPELINE_SHADER_STAGE_CREATE_REQUIRE_FULL_SUBGROUPS_BIT_EXT
Provided by VK_EXT_subgroup_size_control

VK_PIPELINE_STAGE_ACCELERATION_STRUCTURE_BUILD_BIT_KHR

VK_PIPELINE_STAGE_ACCELERATION_STRUCTURE_BUILD_BIT_NV

VK_PIPELINE_STAGE_ALL_COMMANDS_BIT
All stages supported on the queue

VK_PIPELINE_STAGE_ALL_GRAPHICS_BIT
All stages of the graphics pipeline

VK_PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT
After previous commands have completed

VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT
Color attachment writes

VK_PIPELINE_STAGE_COMMAND_PREPROCESS_BIT_NV

VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT
Compute shading

VK_PIPELINE_STAGE_CONDITIONAL_RENDERING_BIT_EXT

VK_PIPELINE_STAGE_DRAW_INDIRECT_BIT
Draw/DispatchIndirect command fetch

VK_PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT
Early fragment (depth and stencil) tests

VK_PIPELINE_STAGE_FRAGMENT_DENSITY_PROCESS_BIT_EXT

VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT
Fragment shading

VK_PIPELINE_STAGE_FRAGMENT_SHADING_RATE_ATTACHMENT_BIT_KHR

VK_PIPELINE_STAGE_GEOMETRY_SHADER_BIT
Geometry shading

VK_PIPELINE_STAGE_HOST_BIT
Indicates host (CPU) is a source/sink of the dependency

VK_PIPELINE_STAGE_LATE_FRAGMENT_TESTS_BIT
Late fragment (depth and stencil) tests

VK_PIPELINE_STAGE_MESH_SHADER_BIT_NV

VK_PIPELINE_STAGE_NONE_KHR

VK_PIPELINE_STAGE_RAY_TRACING_SHADER_BIT_KHR

VK_PIPELINE_STAGE_RAY_TRACING_SHADER_BIT_NV

VK_PIPELINE_STAGE_SHADING_RATE_IMAGE_BIT_NV

VK_PIPELINE_STAGE_TASK_SHADER_BIT_NV

VK_PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT
Tessellation control shading

VK_PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT
Tessellation evaluation shading

VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT
Before subsequent commands are processed

VK_PIPELINE_STAGE_TRANSFER_BIT
Transfer/copy operations

VK_PIPELINE_STAGE_TRANSFORM_FEEDBACK_BIT_EXT

VK_PIPELINE_STAGE_VERTEX_INPUT_BIT
Vertex/index fetch

VK_PIPELINE_STAGE_VERTEX_SHADER_BIT
Vertex shading

VK_QUERY_CONTROL_PRECISE_BIT
Require precise results to be collected by the query

VK_QUERY_PIPELINE_STATISTIC_CLIPPING_INVOCATIONS_BIT
Optional

VK_QUERY_PIPELINE_STATISTIC_CLIPPING_PRIMITIVES_BIT
Optional

VK_QUERY_PIPELINE_STATISTIC_COMPUTE_SHADER_INVOCATIONS_BIT
Optional

VK_QUERY_PIPELINE_STATISTIC_FRAGMENT_SHADER_INVOCATIONS_BIT
Optional

VK_QUERY_PIPELINE_STATISTIC_GEOMETRY_SHADER_INVOCATIONS_BIT
Optional

VK_QUERY_PIPELINE_STATISTIC_GEOMETRY_SHADER_PRIMITIVES_BIT
Optional

VK_QUERY_PIPELINE_STATISTIC_INPUT_ASSEMBLY_PRIMITIVES_BIT
Optional

VK_QUERY_PIPELINE_STATISTIC_INPUT_ASSEMBLY_VERTICES_BIT
Optional

VK_QUERY_PIPELINE_STATISTIC_TESSELLATION_CONTROL_SHADER_PATCHES_BIT
Optional

VK_QUERY_PIPELINE_STATISTIC_TESSELLATION_EVALUATION_SHADER_INVOCATIONS_BIT
Optional

VK_QUERY_PIPELINE_STATISTIC_VERTEX_SHADER_INVOCATIONS_BIT
Optional

VK_QUERY_RESULT_64_BIT
Results of the queries are written to the destination buffer as 64-bit values

VK_QUERY_RESULT_PARTIAL_BIT
Copy the partial results of the query even if the final results are not available

VK_QUERY_RESULT_WAIT_BIT
Results of the queries are waited on before proceeding with the result copy

VK_QUERY_RESULT_WITH_AVAILABILITY_BIT
Besides the results of the query, the availability of the results is also written

VK_QUEUE_COMPUTE_BIT
Queue supports compute operations

VK_QUEUE_GRAPHICS_BIT
Queue supports graphics operations

VK_QUEUE_PROTECTED_BIT
Provided by VK_VERSION_1_1

VK_QUEUE_SPARSE_BINDING_BIT
Queue supports sparse resource memory management operations

VK_QUEUE_TRANSFER_BIT
Queue supports transfer operations

VK_RENDER_PASS_CREATE_TRANSFORM_BIT_QCOM
Provided by VK_QCOM_render_pass_transform

VK_SAMPLER_CREATE_SUBSAMPLED_BIT_EXT
Provided by VK_EXT_fragment_density_map

VK_SAMPLER_CREATE_SUBSAMPLED_COARSE_RECONSTRUCTION_BIT_EXT
Provided by VK_EXT_fragment_density_map

VK_SAMPLE_COUNT_1_BIT
Sample count 1 supported

VK_SAMPLE_COUNT_2_BIT
Sample count 2 supported

VK_SAMPLE_COUNT_4_BIT
Sample count 4 supported

VK_SAMPLE_COUNT_8_BIT
Sample count 8 supported

VK_SAMPLE_COUNT_16_BIT
Sample count 16 supported

VK_SAMPLE_COUNT_32_BIT
Sample count 32 supported

VK_SAMPLE_COUNT_64_BIT
Sample count 64 supported

VK_SHADER_STAGE_ALL

VK_SHADER_STAGE_ALL_GRAPHICS

VK_SHADER_STAGE_ANY_HIT_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_SHADER_STAGE_ANY_HIT_BIT_NV
Provided by VK_NV_ray_tracing

VK_SHADER_STAGE_CALLABLE_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_SHADER_STAGE_CALLABLE_BIT_NV
Provided by VK_NV_ray_tracing

VK_SHADER_STAGE_CLOSEST_HIT_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_SHADER_STAGE_CLOSEST_HIT_BIT_NV
Provided by VK_NV_ray_tracing

VK_SHADER_STAGE_COMPUTE_BIT

VK_SHADER_STAGE_FRAGMENT_BIT

VK_SHADER_STAGE_GEOMETRY_BIT

VK_SHADER_STAGE_INTERSECTION_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_SHADER_STAGE_INTERSECTION_BIT_NV
Provided by VK_NV_ray_tracing

VK_SHADER_STAGE_MESH_BIT_NV
Provided by VK_NV_mesh_shader

VK_SHADER_STAGE_MISS_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_SHADER_STAGE_MISS_BIT_NV
Provided by VK_NV_ray_tracing

VK_SHADER_STAGE_RAYGEN_BIT_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_SHADER_STAGE_RAYGEN_BIT_NV
Provided by VK_NV_ray_tracing

VK_SHADER_STAGE_TASK_BIT_NV
Provided by VK_NV_mesh_shader

VK_SHADER_STAGE_TESSELLATION_CONTROL_BIT

VK_SHADER_STAGE_TESSELLATION_EVALUATION_BIT

VK_SHADER_STAGE_VERTEX_BIT

VK_SPARSE_IMAGE_FORMAT_ALIGNED_MIP_SIZE_BIT
Image requires mip level dimensions to be an integer multiple of the sparse image block dimensions for non-tail mip levels.

VK_SPARSE_IMAGE_FORMAT_NONSTANDARD_BLOCK_SIZE_BIT
Image uses a non-standard sparse image block dimensions

VK_SPARSE_IMAGE_FORMAT_SINGLE_MIPTAIL_BIT
Image uses a single mip tail region for all array layers

VK_SPARSE_MEMORY_BIND_METADATA_BIT
Operation binds resource metadata to memory

VK_STENCIL_FACE_BACK_BIT
Back face

VK_STENCIL_FACE_FRONT_AND_BACK
Front and back faces

VK_STENCIL_FACE_FRONT_BIT
Front face

VK_STENCIL_FRONT_AND_BACK
Alias for backwards compatibility

Type Definitions

VkAccessFlags

VkAttachmentDescriptionFlags

VkBufferCreateFlags

VkBufferUsageFlags

VkBufferViewCreateFlags

VkColorComponentFlags

VkCommandBufferResetFlags

VkCommandBufferUsageFlags

VkCommandPoolCreateFlags

VkCommandPoolResetFlags

VkCullModeFlags

VkDependencyFlags

VkDescriptorPoolCreateFlags

VkDescriptorPoolResetFlags

VkDescriptorSetLayoutCreateFlags

VkDeviceCreateFlags

VkDeviceQueueCreateFlags

VkEventCreateFlags

VkFenceCreateFlags

VkFormatFeatureFlags

VkFramebufferCreateFlags

VkImageAspectFlags

VkImageCreateFlags

VkImageUsageFlags

VkImageViewCreateFlags

VkInstanceCreateFlags

VkMemoryHeapFlags

VkMemoryMapFlags

VkMemoryPropertyFlags

VkPipelineCacheCreateFlags

VkPipelineColorBlendStateCreateFlags

VkPipelineCreateFlags

VkPipelineDepthStencilStateCreateFlags

VkPipelineDynamicStateCreateFlags

VkPipelineInputAssemblyStateCreateFlags

VkPipelineLayoutCreateFlags

VkPipelineMultisampleStateCreateFlags

VkPipelineRasterizationStateCreateFlags

VkPipelineShaderStageCreateFlags

VkPipelineStageFlags

VkPipelineTessellationStateCreateFlags

VkPipelineVertexInputStateCreateFlags

VkPipelineViewportStateCreateFlags

VkQueryControlFlags

VkQueryPipelineStatisticFlags

VkQueryPoolCreateFlags

VkQueryResultFlags

VkQueueFlags

VkRenderPassCreateFlags

VkSampleCountFlags

VkSamplerCreateFlags

VkSemaphoreCreateFlags

VkShaderModuleCreateFlags

VkShaderStageFlags

VkSparseImageFormatFlags

VkSparseMemoryBindFlags

VkStencilFaceFlags

VkSubpassDescriptionFlags