Struct vkvk::fn_managers_rusty::PreInstanceFnsRusty

pub struct PreInstanceFnsRusty(pub PreInstanceFns);

Implementations

impl PreInstanceFnsRusty

pub fn enumerate_instance_layer_properties(
&self
) -> Result<Vec<VkLayerProperties>, VkResult>

Gets info about the available layers.

pub fn enumerate_instance_extension_properties(
&self,
layer_name: Option<&[u8; 256]>
) -> Result<Vec<VkExtensionProperties>, VkResult>

Gets info about the extensions available per layer.

Extensions are specific to a given layer, or to the “base” instance.

Use Some(layer_name) to get extension info for the named layer (layer names come from enumerate_instance_layer_properties)

Use None to get extension info about the base instance.

pub fn create_instance_simple(
&self,
name: &str,
api: VulkanVersion,
layers: &[VkLayerProperties],
extensions: &[VkExtensionProperties]
) -> Result<VkInstance, VkResult>