Struct vkvk::structures::VkPhysicalDeviceLimits

#[repr(C)]pub struct VkPhysicalDeviceLimits {
    pub maxImageDimension1D: u32,
    pub maxImageDimension2D: u32,
    pub maxImageDimension3D: u32,
    pub maxImageDimensionCube: u32,
    pub maxImageArrayLayers: u32,
    pub maxTexelBufferElements: u32,
    pub maxUniformBufferRange: u32,
    pub maxStorageBufferRange: u32,
    pub maxPushConstantsSize: u32,
    pub maxMemoryAllocationCount: u32,
    pub maxSamplerAllocationCount: u32,
    pub bufferImageGranularity: VkDeviceSize,
    pub sparseAddressSpaceSize: VkDeviceSize,
    pub maxBoundDescriptorSets: u32,
    pub maxPerStageDescriptorSamplers: u32,
    pub maxPerStageDescriptorUniformBuffers: u32,
    pub maxPerStageDescriptorStorageBuffers: u32,
    pub maxPerStageDescriptorSampledImages: u32,
    pub maxPerStageDescriptorStorageImages: u32,
    pub maxPerStageDescriptorInputAttachments: u32,
    pub maxPerStageResources: u32,
    pub maxDescriptorSetSamplers: u32,
    pub maxDescriptorSetUniformBuffers: u32,
    pub maxDescriptorSetUniformBuffersDynamic: u32,
    pub maxDescriptorSetStorageBuffers: u32,
    pub maxDescriptorSetStorageBuffersDynamic: u32,
    pub maxDescriptorSetSampledImages: u32,
    pub maxDescriptorSetStorageImages: u32,
    pub maxDescriptorSetInputAttachments: u32,
    pub maxVertexInputAttributes: u32,
    pub maxVertexInputBindings: u32,
    pub maxVertexInputAttributeOffset: u32,
    pub maxVertexInputBindingStride: u32,
    pub maxVertexOutputComponents: u32,
    pub maxTessellationGenerationLevel: u32,
    pub maxTessellationPatchSize: u32,
    pub maxTessellationControlPerVertexInputComponents: u32,
    pub maxTessellationControlPerVertexOutputComponents: u32,
    pub maxTessellationControlPerPatchOutputComponents: u32,
    pub maxTessellationControlTotalOutputComponents: u32,
    pub maxTessellationEvaluationInputComponents: u32,
    pub maxTessellationEvaluationOutputComponents: u32,
    pub maxGeometryShaderInvocations: u32,
    pub maxGeometryInputComponents: u32,
    pub maxGeometryOutputComponents: u32,
    pub maxGeometryOutputVertices: u32,
    pub maxGeometryTotalOutputComponents: u32,
    pub maxFragmentInputComponents: u32,
    pub maxFragmentOutputAttachments: u32,
    pub maxFragmentDualSrcAttachments: u32,
    pub maxFragmentCombinedOutputResources: u32,
    pub maxComputeSharedMemorySize: u32,
    pub maxComputeWorkGroupCount: [u32; 3],
    pub maxComputeWorkGroupInvocations: u32,
    pub maxComputeWorkGroupSize: [u32; 3],
    pub subPixelPrecisionBits: u32,
    pub subTexelPrecisionBits: u32,
    pub mipmapPrecisionBits: u32,
    pub maxDrawIndexedIndexValue: u32,
    pub maxDrawIndirectCount: u32,
    pub maxSamplerLodBias: f32,
    pub maxSamplerAnisotropy: f32,
    pub maxViewports: u32,
    pub maxViewportDimensions: [u32; 2],
    pub viewportBoundsRange: [f32; 2],
    pub viewportSubPixelBits: u32,
    pub minMemoryMapAlignment: usize,
    pub minTexelBufferOffsetAlignment: VkDeviceSize,
    pub minUniformBufferOffsetAlignment: VkDeviceSize,
    pub minStorageBufferOffsetAlignment: VkDeviceSize,
    pub minTexelOffset: i32,
    pub maxTexelOffset: u32,
    pub minTexelGatherOffset: i32,
    pub maxTexelGatherOffset: u32,
    pub minInterpolationOffset: f32,
    pub maxInterpolationOffset: f32,
    pub subPixelInterpolationOffsetBits: u32,
    pub maxFramebufferWidth: u32,
    pub maxFramebufferHeight: u32,
    pub maxFramebufferLayers: u32,
    pub framebufferColorSampleCounts: VkSampleCountFlags,
    pub framebufferDepthSampleCounts: VkSampleCountFlags,
    pub framebufferStencilSampleCounts: VkSampleCountFlags,
    pub framebufferNoAttachmentsSampleCounts: VkSampleCountFlags,
    pub maxColorAttachments: u32,
    pub sampledImageColorSampleCounts: VkSampleCountFlags,
    pub sampledImageIntegerSampleCounts: VkSampleCountFlags,
    pub sampledImageDepthSampleCounts: VkSampleCountFlags,
    pub sampledImageStencilSampleCounts: VkSampleCountFlags,
    pub storageImageSampleCounts: VkSampleCountFlags,
    pub maxSampleMaskWords: u32,
    pub timestampComputeAndGraphics: VkBool32,
    pub timestampPeriod: f32,
    pub maxClipDistances: u32,
    pub maxCullDistances: u32,
    pub maxCombinedClipAndCullDistances: u32,
    pub discreteQueuePriorities: u32,
    pub pointSizeRange: [f32; 2],
    pub lineWidthRange: [f32; 2],
    pub pointSizeGranularity: f32,
    pub lineWidthGranularity: f32,
    pub strictLines: VkBool32,
    pub standardSampleLocations: VkBool32,
    pub optimalBufferCopyOffsetAlignment: VkDeviceSize,
    pub optimalBufferCopyRowPitchAlignment: VkDeviceSize,
    pub nonCoherentAtomSize: VkDeviceSize,
}

VkPhysicalDeviceLimits

Fields
maxImageDimension1D: u32

max 1D image dimension

maxImageDimension2D: u32

max 2D image dimension

maxImageDimension3D: u32

max 3D image dimension

maxImageDimensionCube: u32

max cubemap image dimension

maxImageArrayLayers: u32

max layers for image arrays

maxTexelBufferElements: u32

max texel buffer size (fstexels)

maxUniformBufferRange: u32

max uniform buffer range (bytes)

maxStorageBufferRange: u32

max storage buffer range (bytes)

maxPushConstantsSize: u32

max size of the push constants pool (bytes)

maxMemoryAllocationCount: u32

max number of device memory allocations supported

maxSamplerAllocationCount: u32

max number of samplers that can be allocated on a device

bufferImageGranularity: VkDeviceSize

Granularity (in bytes) at which buffers and images can be bound to adjacent memory for simultaneous usage

sparseAddressSpaceSize: VkDeviceSize

Total address space available for sparse allocations (bytes)

maxBoundDescriptorSets: u32

max number of descriptors sets that can be bound to a pipeline

maxPerStageDescriptorSamplers: u32

max number of samplers allowed per-stage in a descriptor set

maxPerStageDescriptorUniformBuffers: u32

max number of uniform buffers allowed per-stage in a descriptor set

maxPerStageDescriptorStorageBuffers: u32

max number of storage buffers allowed per-stage in a descriptor set

maxPerStageDescriptorSampledImages: u32

max number of sampled images allowed per-stage in a descriptor set

maxPerStageDescriptorStorageImages: u32

max number of storage images allowed per-stage in a descriptor set

maxPerStageDescriptorInputAttachments: u32

max number of input attachments allowed per-stage in a descriptor set

maxPerStageResources: u32

max number of resources allowed by a single stage

maxDescriptorSetSamplers: u32

max number of samplers allowed in all stages in a descriptor set

maxDescriptorSetUniformBuffers: u32

max number of uniform buffers allowed in all stages in a descriptor set

maxDescriptorSetUniformBuffersDynamic: u32

max number of dynamic uniform buffers allowed in all stages in a descriptor set

maxDescriptorSetStorageBuffers: u32

max number of storage buffers allowed in all stages in a descriptor set

maxDescriptorSetStorageBuffersDynamic: u32

max number of dynamic storage buffers allowed in all stages in a descriptor set

maxDescriptorSetSampledImages: u32

max number of sampled images allowed in all stages in a descriptor set

maxDescriptorSetStorageImages: u32

max number of storage images allowed in all stages in a descriptor set

maxDescriptorSetInputAttachments: u32

max number of input attachments allowed in all stages in a descriptor set

maxVertexInputAttributes: u32

max number of vertex input attribute slots

maxVertexInputBindings: u32

max number of vertex input binding slots

maxVertexInputAttributeOffset: u32

max vertex input attribute offset added to vertex buffer offset

maxVertexInputBindingStride: u32

max vertex input binding stride

maxVertexOutputComponents: u32

max number of output components written by vertex shader

maxTessellationGenerationLevel: u32

max level supported by tessellation primitive generator

maxTessellationPatchSize: u32

max patch size (vertices)

maxTessellationControlPerVertexInputComponents: u32

max number of input components per-vertex in TCS

maxTessellationControlPerVertexOutputComponents: u32

max number of output components per-vertex in TCS

maxTessellationControlPerPatchOutputComponents: u32

max number of output components per-patch in TCS

maxTessellationControlTotalOutputComponents: u32

max total number of per-vertex and per-patch output components in TCS

maxTessellationEvaluationInputComponents: u32

max number of input components per vertex in TES

maxTessellationEvaluationOutputComponents: u32

max number of output components per vertex in TES

maxGeometryShaderInvocations: u32

max invocation count supported in geometry shader

maxGeometryInputComponents: u32

max number of input components read in geometry stage

maxGeometryOutputComponents: u32

max number of output components written in geometry stage

maxGeometryOutputVertices: u32

max number of vertices that can be emitted in geometry stage

maxGeometryTotalOutputComponents: u32

max total number of components (all vertices) written in geometry stage

maxFragmentInputComponents: u32

max number of input components read in fragment stage

maxFragmentOutputAttachments: u32

max number of output attachments written in fragment stage

maxFragmentDualSrcAttachments: u32

max number of output attachments written when using dual source blending

maxFragmentCombinedOutputResources: u32

max total number of storage buffers, storage images and output buffers

maxComputeSharedMemorySize: u32

max total storage size of work group local storage (bytes)

maxComputeWorkGroupCount: [u32; 3]

max num of compute work groups that may be dispatched by a single command (x,y,z)

maxComputeWorkGroupInvocations: u32

max total compute invocations in a single local work group

maxComputeWorkGroupSize: [u32; 3]

max local size of a compute work group (x,y,z)

subPixelPrecisionBits: u32

number bits of subpixel precision in screen x and y

subTexelPrecisionBits: u32

number bits of precision for selecting texel weights

mipmapPrecisionBits: u32

number bits of precision for selecting mipmap weights

maxDrawIndexedIndexValue: u32

max index value for indexed draw calls (for 32-bit indices)

maxDrawIndirectCount: u32

max draw count for indirect draw calls

maxSamplerLodBias: f32

max absolute sampler LOD bias

maxSamplerAnisotropy: f32

max degree of sampler anisotropy

maxViewports: u32

max number of active viewports

maxViewportDimensions: [u32; 2]

max viewport dimensions (x,y)

viewportBoundsRange: [f32; 2]

viewport bounds range (min,max)

viewportSubPixelBits: u32

number bits of subpixel precision for viewport

minMemoryMapAlignment: usize

min required alignment of pointers returned by MapMemory (bytes)

minTexelBufferOffsetAlignment: VkDeviceSize

min required alignment for texel buffer offsets (bytes)

minUniformBufferOffsetAlignment: VkDeviceSize

min required alignment for uniform buffer sizes and offsets (bytes)

minStorageBufferOffsetAlignment: VkDeviceSize

min required alignment for storage buffer offsets (bytes)

minTexelOffset: i32

min texel offset for OpTextureSampleOffset

maxTexelOffset: u32

max texel offset for OpTextureSampleOffset

minTexelGatherOffset: i32

min texel offset for OpTextureGatherOffset

maxTexelGatherOffset: u32

max texel offset for OpTextureGatherOffset

minInterpolationOffset: f32

furthest negative offset for interpolateAtOffset

maxInterpolationOffset: f32

furthest positive offset for interpolateAtOffset

subPixelInterpolationOffsetBits: u32

number of subpixel bits for interpolateAtOffset

maxFramebufferWidth: u32

max width for a framebuffer

maxFramebufferHeight: u32

max height for a framebuffer

maxFramebufferLayers: u32

max layer count for a layered framebuffer

framebufferColorSampleCounts: VkSampleCountFlags

supported color sample counts for a framebuffer

Optional: true

framebufferDepthSampleCounts: VkSampleCountFlags

supported depth sample counts for a framebuffer

Optional: true

framebufferStencilSampleCounts: VkSampleCountFlags

supported stencil sample counts for a framebuffer

Optional: true

framebufferNoAttachmentsSampleCounts: VkSampleCountFlags

supported sample counts for a subpass which uses no attachments

Optional: true

maxColorAttachments: u32

max number of color attachments per subpass

sampledImageColorSampleCounts: VkSampleCountFlags

supported color sample counts for a non-integer sampled image

Optional: true

sampledImageIntegerSampleCounts: VkSampleCountFlags

supported sample counts for an integer image

Optional: true

sampledImageDepthSampleCounts: VkSampleCountFlags

supported depth sample counts for a sampled image

Optional: true

sampledImageStencilSampleCounts: VkSampleCountFlags

supported stencil sample counts for a sampled image

Optional: true

storageImageSampleCounts: VkSampleCountFlags

supported sample counts for a storage image

Optional: true

maxSampleMaskWords: u32

max number of sample mask words

timestampComputeAndGraphics: VkBool32

timestamps on graphics and compute queues

timestampPeriod: f32

number of nanoseconds it takes for timestamp query value to increment by 1

maxClipDistances: u32

max number of clip distances

maxCullDistances: u32

max number of cull distances

maxCombinedClipAndCullDistances: u32

max combined number of user clipping

discreteQueuePriorities: u32

distinct queue priorities available

pointSizeRange: [f32; 2]

range (min,max) of supported point sizes

lineWidthRange: [f32; 2]

range (min,max) of supported line widths

pointSizeGranularity: f32

granularity of supported point sizes

lineWidthGranularity: f32

granularity of supported line widths

strictLines: VkBool32

line rasterization follows preferred rules

standardSampleLocations: VkBool32

supports standard sample locations for all supported sample counts

optimalBufferCopyOffsetAlignment: VkDeviceSize

optimal offset of buffer copies

optimalBufferCopyRowPitchAlignment: VkDeviceSize

optimal pitch of buffer copies

nonCoherentAtomSize: VkDeviceSize

minimum size and alignment for non-coherent host-mapped device memory access