Module vkvk::vk_result

Structs

VkResult
API result codes.

Constants

VK_ERROR_DEVICE_LOST
The logical device has been lost.

VK_ERROR_EXTENSION_NOT_PRESENT
Extension specified does not exist

VK_ERROR_FEATURE_NOT_PRESENT
Requested feature is not available on this device

VK_ERROR_FORMAT_NOT_SUPPORTED
Requested format is not supported on this device

VK_ERROR_FRAGMENTATION
Provided by VK_VERSION_1_2

VK_ERROR_FRAGMENTATION_EXT
Provided by VK_EXT_descriptor_indexing

VK_ERROR_FRAGMENTED_POOL
A requested pool allocation has failed due to fragmentation of the pool’s memory

VK_ERROR_FULL_SCREEN_EXCLUSIVE_MODE_LOST_EXT
Provided by VK_EXT_full_screen_exclusive

VK_ERROR_INCOMPATIBLE_DISPLAY_KHR
Provided by VK_KHR_display_swapchain

VK_ERROR_INCOMPATIBLE_DRIVER
Unable to find a Vulkan driver

VK_ERROR_INITIALIZATION_FAILED
Initialization of a object has failed

VK_ERROR_INVALID_DEVICE_ADDRESS_EXT
Provided by VK_EXT_buffer_device_address

VK_ERROR_INVALID_DRM_FORMAT_MODIFIER_PLANE_LAYOUT_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_ERROR_INVALID_EXTERNAL_HANDLE
Provided by VK_VERSION_1_1

VK_ERROR_INVALID_EXTERNAL_HANDLE_KHR
Provided by VK_KHR_external_memory

VK_ERROR_INVALID_OPAQUE_CAPTURE_ADDRESS
Provided by VK_VERSION_1_2

VK_ERROR_INVALID_OPAQUE_CAPTURE_ADDRESS_KHR
Provided by VK_KHR_buffer_device_address

VK_ERROR_INVALID_SHADER_NV
Provided by VK_NV_glsl_shader

VK_ERROR_LAYER_NOT_PRESENT
Layer specified does not exist

VK_ERROR_MEMORY_MAP_FAILED
Mapping of a memory object has failed

VK_ERROR_NATIVE_WINDOW_IN_USE_KHR
Provided by VK_KHR_surface

VK_ERROR_NOT_PERMITTED_EXT
Provided by VK_EXT_global_priority

VK_ERROR_OUT_OF_DATE_KHR
Provided by VK_KHR_swapchain

VK_ERROR_OUT_OF_DEVICE_MEMORY
A device memory allocation has failed

VK_ERROR_OUT_OF_HOST_MEMORY
A host memory allocation has failed

VK_ERROR_OUT_OF_POOL_MEMORY

VK_ERROR_OUT_OF_POOL_MEMORY_KHR
Provided by VK_KHR_maintenance1

VK_ERROR_PIPELINE_COMPILE_REQUIRED_EXT

VK_ERROR_SURFACE_LOST_KHR
Provided by VK_KHR_surface

VK_ERROR_TOO_MANY_OBJECTS
Too many objects of the type have already been created

VK_ERROR_UNKNOWN
An unknown error has occurred, due to an implementation or application bug

VK_ERROR_VALIDATION_FAILED_EXT
Provided by VK_EXT_debug_report

VK_EVENT_RESET
An event is unsignaled

VK_EVENT_SET
An event is signaled

VK_INCOMPLETE
A return array was too small for the result

VK_NOT_READY
A fence or query has not yet completed

VK_OPERATION_DEFERRED_KHR
Provided by VK_KHR_deferred_host_operations

VK_OPERATION_NOT_DEFERRED_KHR
Provided by VK_KHR_deferred_host_operations

VK_PIPELINE_COMPILE_REQUIRED_EXT
Provided by VK_EXT_pipeline_creation_cache_control

VK_SUBOPTIMAL_KHR
Provided by VK_KHR_swapchain

VK_SUCCESS
Command completed successfully

VK_THREAD_DONE_KHR
Provided by VK_KHR_deferred_host_operations

VK_THREAD_IDLE_KHR
Provided by VK_KHR_deferred_host_operations

VK_TIMEOUT
A wait operation has not completed in the specified time