Module vkvk::vk_structure_type

Structs

VkStructureType
Structure type enumerant.

Constants

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_BUILD_GEOMETRY_INFO_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_BUILD_SIZES_INFO_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_CREATE_INFO_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_CREATE_INFO_NV
Provided by VK_NV_ray_tracing

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_DEVICE_ADDRESS_INFO_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_GEOMETRY_AABBS_DATA_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_GEOMETRY_INSTANCES_DATA_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_GEOMETRY_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_GEOMETRY_TRIANGLES_DATA_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_INFO_NV
Provided by VK_NV_ray_tracing

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_MEMORY_REQUIREMENTS_INFO_NV
Provided by VK_NV_ray_tracing

VK_STRUCTURE_TYPE_ACCELERATION_STRUCTURE_VERSION_INFO_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_ACQUIRE_NEXT_IMAGE_INFO_KHR
Provided by VK_KHR_swapchain with VK_VERSION_1_1, VK_KHR_device_group with VK_KHR_swapchain

VK_STRUCTURE_TYPE_ACQUIRE_PROFILING_LOCK_INFO_KHR
Provided by VK_KHR_performance_query

VK_STRUCTURE_TYPE_ANDROID_HARDWARE_BUFFER_FORMAT_PROPERTIES_ANDROID
Provided by VK_ANDROID_external_memory_android_hardware_buffer

VK_STRUCTURE_TYPE_ANDROID_HARDWARE_BUFFER_PROPERTIES_ANDROID
Provided by VK_ANDROID_external_memory_android_hardware_buffer

VK_STRUCTURE_TYPE_ANDROID_HARDWARE_BUFFER_USAGE_ANDROID
Provided by VK_ANDROID_external_memory_android_hardware_buffer

VK_STRUCTURE_TYPE_ANDROID_SURFACE_CREATE_INFO_KHR
Provided by VK_KHR_android_surface

VK_STRUCTURE_TYPE_APPLICATION_INFO

VK_STRUCTURE_TYPE_ATTACHMENT_DESCRIPTION_2
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_ATTACHMENT_DESCRIPTION_2_KHR
Provided by VK_KHR_create_renderpass2

VK_STRUCTURE_TYPE_ATTACHMENT_DESCRIPTION_STENCIL_LAYOUT
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_ATTACHMENT_DESCRIPTION_STENCIL_LAYOUT_KHR
Provided by VK_KHR_separate_depth_stencil_layouts

VK_STRUCTURE_TYPE_ATTACHMENT_REFERENCE_2
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_ATTACHMENT_REFERENCE_2_KHR
Provided by VK_KHR_create_renderpass2

VK_STRUCTURE_TYPE_ATTACHMENT_REFERENCE_STENCIL_LAYOUT
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_ATTACHMENT_REFERENCE_STENCIL_LAYOUT_KHR
Provided by VK_KHR_separate_depth_stencil_layouts

VK_STRUCTURE_TYPE_BIND_ACCELERATION_STRUCTURE_MEMORY_INFO_NV
Provided by VK_NV_ray_tracing

VK_STRUCTURE_TYPE_BIND_BUFFER_MEMORY_DEVICE_GROUP_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_BIND_BUFFER_MEMORY_DEVICE_GROUP_INFO_KHR
Provided by VK_KHR_device_group with VK_KHR_bind_memory2

VK_STRUCTURE_TYPE_BIND_BUFFER_MEMORY_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_BIND_BUFFER_MEMORY_INFO_KHR
Provided by VK_KHR_bind_memory2

VK_STRUCTURE_TYPE_BIND_IMAGE_MEMORY_DEVICE_GROUP_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_BIND_IMAGE_MEMORY_DEVICE_GROUP_INFO_KHR
Provided by VK_KHR_device_group with VK_KHR_bind_memory2

VK_STRUCTURE_TYPE_BIND_IMAGE_MEMORY_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_BIND_IMAGE_MEMORY_INFO_KHR
Provided by VK_KHR_bind_memory2

VK_STRUCTURE_TYPE_BIND_IMAGE_MEMORY_SWAPCHAIN_INFO_KHR
Provided by VK_KHR_swapchain with VK_VERSION_1_1, VK_KHR_device_group with VK_KHR_swapchain

VK_STRUCTURE_TYPE_BIND_IMAGE_PLANE_MEMORY_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_BIND_IMAGE_PLANE_MEMORY_INFO_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_STRUCTURE_TYPE_BIND_SPARSE_INFO

VK_STRUCTURE_TYPE_BLIT_IMAGE_INFO_2_KHR
Provided by VK_KHR_copy_commands2

VK_STRUCTURE_TYPE_BUFFER_COPY_2_KHR
Provided by VK_KHR_copy_commands2

VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO

VK_STRUCTURE_TYPE_BUFFER_DEVICE_ADDRESS_CREATE_INFO_EXT
Provided by VK_EXT_buffer_device_address

VK_STRUCTURE_TYPE_BUFFER_DEVICE_ADDRESS_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_BUFFER_DEVICE_ADDRESS_INFO_EXT
Provided by VK_EXT_buffer_device_address

VK_STRUCTURE_TYPE_BUFFER_DEVICE_ADDRESS_INFO_KHR
Provided by VK_KHR_buffer_device_address

VK_STRUCTURE_TYPE_BUFFER_IMAGE_COPY_2_KHR
Provided by VK_KHR_copy_commands2

VK_STRUCTURE_TYPE_BUFFER_MEMORY_BARRIER

VK_STRUCTURE_TYPE_BUFFER_MEMORY_REQUIREMENTS_INFO_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_BUFFER_MEMORY_REQUIREMENTS_INFO_2_KHR
Provided by VK_KHR_get_memory_requirements2

VK_STRUCTURE_TYPE_BUFFER_OPAQUE_CAPTURE_ADDRESS_CREATE_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_BUFFER_OPAQUE_CAPTURE_ADDRESS_CREATE_INFO_KHR
Provided by VK_KHR_buffer_device_address

VK_STRUCTURE_TYPE_BUFFER_VIEW_CREATE_INFO

VK_STRUCTURE_TYPE_CALIBRATED_TIMESTAMP_INFO_EXT
Provided by VK_EXT_calibrated_timestamps

VK_STRUCTURE_TYPE_CHECKPOINT_DATA_NV
Provided by VK_NV_device_diagnostic_checkpoints

VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO

VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO

VK_STRUCTURE_TYPE_COMMAND_BUFFER_INHERITANCE_CONDITIONAL_RENDERING_INFO_EXT
Provided by VK_EXT_conditional_rendering

VK_STRUCTURE_TYPE_COMMAND_BUFFER_INHERITANCE_INFO

VK_STRUCTURE_TYPE_COMMAND_BUFFER_INHERITANCE_RENDER_PASS_TRANSFORM_INFO_QCOM
Provided by VK_QCOM_render_pass_transform

VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO

VK_STRUCTURE_TYPE_COMPUTE_PIPELINE_CREATE_INFO

VK_STRUCTURE_TYPE_CONDITIONAL_RENDERING_BEGIN_INFO_EXT
Provided by VK_EXT_conditional_rendering

VK_STRUCTURE_TYPE_COOPERATIVE_MATRIX_PROPERTIES_NV
Provided by VK_NV_cooperative_matrix

VK_STRUCTURE_TYPE_COPY_ACCELERATION_STRUCTURE_INFO_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_COPY_ACCELERATION_STRUCTURE_TO_MEMORY_INFO_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_COPY_BUFFER_INFO_2_KHR
Provided by VK_KHR_copy_commands2

VK_STRUCTURE_TYPE_COPY_BUFFER_TO_IMAGE_INFO_2_KHR
Provided by VK_KHR_copy_commands2

VK_STRUCTURE_TYPE_COPY_COMMAND_TRANSFORM_INFO_QCOM
Provided by VK_QCOM_rotated_copy_commands

VK_STRUCTURE_TYPE_COPY_DESCRIPTOR_SET

VK_STRUCTURE_TYPE_COPY_IMAGE_INFO_2_KHR
Provided by VK_KHR_copy_commands2

VK_STRUCTURE_TYPE_COPY_IMAGE_TO_BUFFER_INFO_2_KHR
Provided by VK_KHR_copy_commands2

VK_STRUCTURE_TYPE_COPY_MEMORY_TO_ACCELERATION_STRUCTURE_INFO_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_D3D12_FENCE_SUBMIT_INFO_KHR
Provided by VK_KHR_external_semaphore_win32

VK_STRUCTURE_TYPE_DEBUG_MARKER_MARKER_INFO_EXT
Provided by VK_EXT_debug_marker

VK_STRUCTURE_TYPE_DEBUG_MARKER_OBJECT_NAME_INFO_EXT
Provided by VK_EXT_debug_marker

VK_STRUCTURE_TYPE_DEBUG_MARKER_OBJECT_TAG_INFO_EXT
Provided by VK_EXT_debug_marker

VK_STRUCTURE_TYPE_DEBUG_REPORT_CALLBACK_CREATE_INFO_EXT
Provided by VK_EXT_debug_report

VK_STRUCTURE_TYPE_DEBUG_REPORT_CREATE_INFO_EXT

VK_STRUCTURE_TYPE_DEBUG_UTILS_LABEL_EXT
Provided by VK_EXT_debug_utils

VK_STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CALLBACK_DATA_EXT
Provided by VK_EXT_debug_utils

VK_STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CREATE_INFO_EXT
Provided by VK_EXT_debug_utils

VK_STRUCTURE_TYPE_DEBUG_UTILS_OBJECT_NAME_INFO_EXT
Provided by VK_EXT_debug_utils

VK_STRUCTURE_TYPE_DEBUG_UTILS_OBJECT_TAG_INFO_EXT
Provided by VK_EXT_debug_utils

VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_BUFFER_CREATE_INFO_NV
Provided by VK_NV_dedicated_allocation

VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_IMAGE_CREATE_INFO_NV
Provided by VK_NV_dedicated_allocation

VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_MEMORY_ALLOCATE_INFO_NV
Provided by VK_NV_dedicated_allocation

VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO

VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_INLINE_UNIFORM_BLOCK_CREATE_INFO_EXT
Provided by VK_EXT_inline_uniform_block

VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO

VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_BINDING_FLAGS_CREATE_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_BINDING_FLAGS_CREATE_INFO_EXT
Provided by VK_EXT_descriptor_indexing

VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO

VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_SUPPORT
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_SUPPORT_KHR
Provided by VK_KHR_maintenance3

VK_STRUCTURE_TYPE_DESCRIPTOR_SET_VARIABLE_DESCRIPTOR_COUNT_ALLOCATE_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_DESCRIPTOR_SET_VARIABLE_DESCRIPTOR_COUNT_ALLOCATE_INFO_EXT
Provided by VK_EXT_descriptor_indexing

VK_STRUCTURE_TYPE_DESCRIPTOR_SET_VARIABLE_DESCRIPTOR_COUNT_LAYOUT_SUPPORT
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_DESCRIPTOR_SET_VARIABLE_DESCRIPTOR_COUNT_LAYOUT_SUPPORT_EXT
Provided by VK_EXT_descriptor_indexing

VK_STRUCTURE_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_CREATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_CREATE_INFO_KHR
Provided by VK_KHR_descriptor_update_template

VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO

VK_STRUCTURE_TYPE_DEVICE_DEVICE_MEMORY_REPORT_CREATE_INFO_EXT
Provided by VK_EXT_device_memory_report

VK_STRUCTURE_TYPE_DEVICE_DIAGNOSTICS_CONFIG_CREATE_INFO_NV
Provided by VK_NV_device_diagnostics_config

VK_STRUCTURE_TYPE_DEVICE_EVENT_INFO_EXT
Provided by VK_EXT_display_control

VK_STRUCTURE_TYPE_DEVICE_GROUP_BIND_SPARSE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_DEVICE_GROUP_BIND_SPARSE_INFO_KHR
Provided by VK_KHR_device_group

VK_STRUCTURE_TYPE_DEVICE_GROUP_COMMAND_BUFFER_BEGIN_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_DEVICE_GROUP_COMMAND_BUFFER_BEGIN_INFO_KHR
Provided by VK_KHR_device_group

VK_STRUCTURE_TYPE_DEVICE_GROUP_DEVICE_CREATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_DEVICE_GROUP_DEVICE_CREATE_INFO_KHR
Provided by VK_KHR_device_group_creation

VK_STRUCTURE_TYPE_DEVICE_GROUP_PRESENT_CAPABILITIES_KHR
Provided by VK_KHR_swapchain with VK_VERSION_1_1, VK_KHR_device_group with VK_KHR_surface

VK_STRUCTURE_TYPE_DEVICE_GROUP_PRESENT_INFO_KHR
Provided by VK_KHR_swapchain with VK_VERSION_1_1, VK_KHR_device_group with VK_KHR_swapchain

VK_STRUCTURE_TYPE_DEVICE_GROUP_RENDER_PASS_BEGIN_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_DEVICE_GROUP_RENDER_PASS_BEGIN_INFO_KHR
Provided by VK_KHR_device_group

VK_STRUCTURE_TYPE_DEVICE_GROUP_SUBMIT_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_DEVICE_GROUP_SUBMIT_INFO_KHR
Provided by VK_KHR_device_group

VK_STRUCTURE_TYPE_DEVICE_GROUP_SWAPCHAIN_CREATE_INFO_KHR
Provided by VK_KHR_swapchain with VK_VERSION_1_1, VK_KHR_device_group with VK_KHR_swapchain

VK_STRUCTURE_TYPE_DEVICE_MEMORY_OPAQUE_CAPTURE_ADDRESS_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_DEVICE_MEMORY_OPAQUE_CAPTURE_ADDRESS_INFO_KHR
Provided by VK_KHR_buffer_device_address

VK_STRUCTURE_TYPE_DEVICE_MEMORY_OVERALLOCATION_CREATE_INFO_AMD
Provided by VK_AMD_memory_overallocation_behavior

VK_STRUCTURE_TYPE_DEVICE_MEMORY_REPORT_CALLBACK_DATA_EXT
Provided by VK_EXT_device_memory_report

VK_STRUCTURE_TYPE_DEVICE_PRIVATE_DATA_CREATE_INFO_EXT
Provided by VK_EXT_private_data

VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO

VK_STRUCTURE_TYPE_DEVICE_QUEUE_GLOBAL_PRIORITY_CREATE_INFO_EXT
Provided by VK_EXT_global_priority

VK_STRUCTURE_TYPE_DEVICE_QUEUE_INFO_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_DIRECTFB_SURFACE_CREATE_INFO_EXT
Provided by VK_EXT_directfb_surface

VK_STRUCTURE_TYPE_DISPLAY_EVENT_INFO_EXT
Provided by VK_EXT_display_control

VK_STRUCTURE_TYPE_DISPLAY_MODE_CREATE_INFO_KHR
Provided by VK_KHR_display

VK_STRUCTURE_TYPE_DISPLAY_MODE_PROPERTIES_2_KHR
Provided by VK_KHR_get_display_properties2

VK_STRUCTURE_TYPE_DISPLAY_NATIVE_HDR_SURFACE_CAPABILITIES_AMD
Provided by VK_AMD_display_native_hdr

VK_STRUCTURE_TYPE_DISPLAY_PLANE_CAPABILITIES_2_KHR
Provided by VK_KHR_get_display_properties2

VK_STRUCTURE_TYPE_DISPLAY_PLANE_INFO_2_KHR
Provided by VK_KHR_get_display_properties2

VK_STRUCTURE_TYPE_DISPLAY_PLANE_PROPERTIES_2_KHR
Provided by VK_KHR_get_display_properties2

VK_STRUCTURE_TYPE_DISPLAY_POWER_INFO_EXT
Provided by VK_EXT_display_control

VK_STRUCTURE_TYPE_DISPLAY_PRESENT_INFO_KHR
Provided by VK_KHR_display_swapchain

VK_STRUCTURE_TYPE_DISPLAY_PROPERTIES_2_KHR
Provided by VK_KHR_get_display_properties2

VK_STRUCTURE_TYPE_DISPLAY_SURFACE_CREATE_INFO_KHR
Provided by VK_KHR_display

VK_STRUCTURE_TYPE_DRM_FORMAT_MODIFIER_PROPERTIES_LIST_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_STRUCTURE_TYPE_EVENT_CREATE_INFO

VK_STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO_KHR
Provided by VK_KHR_external_fence

VK_STRUCTURE_TYPE_EXPORT_FENCE_WIN32_HANDLE_INFO_KHR
Provided by VK_KHR_external_fence_win32

VK_STRUCTURE_TYPE_EXPORT_MEMORY_ALLOCATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_EXPORT_MEMORY_ALLOCATE_INFO_KHR
Provided by VK_KHR_external_memory

VK_STRUCTURE_TYPE_EXPORT_MEMORY_ALLOCATE_INFO_NV
Provided by VK_NV_external_memory

VK_STRUCTURE_TYPE_EXPORT_MEMORY_WIN32_HANDLE_INFO_KHR
Provided by VK_KHR_external_memory_win32

VK_STRUCTURE_TYPE_EXPORT_MEMORY_WIN32_HANDLE_INFO_NV
Provided by VK_NV_external_memory_win32

VK_STRUCTURE_TYPE_EXPORT_SEMAPHORE_CREATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_EXPORT_SEMAPHORE_CREATE_INFO_KHR
Provided by VK_KHR_external_semaphore

VK_STRUCTURE_TYPE_EXPORT_SEMAPHORE_WIN32_HANDLE_INFO_KHR
Provided by VK_KHR_external_semaphore_win32

VK_STRUCTURE_TYPE_EXTERNAL_BUFFER_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_EXTERNAL_BUFFER_PROPERTIES_KHR
Provided by VK_KHR_external_memory_capabilities

VK_STRUCTURE_TYPE_EXTERNAL_FENCE_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_EXTERNAL_FENCE_PROPERTIES_KHR
Provided by VK_KHR_external_fence_capabilities

VK_STRUCTURE_TYPE_EXTERNAL_FORMAT_ANDROID
Provided by VK_ANDROID_external_memory_android_hardware_buffer

VK_STRUCTURE_TYPE_EXTERNAL_IMAGE_FORMAT_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_EXTERNAL_IMAGE_FORMAT_PROPERTIES_KHR
Provided by VK_KHR_external_memory_capabilities

VK_STRUCTURE_TYPE_EXTERNAL_MEMORY_BUFFER_CREATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_EXTERNAL_MEMORY_BUFFER_CREATE_INFO_KHR
Provided by VK_KHR_external_memory

VK_STRUCTURE_TYPE_EXTERNAL_MEMORY_IMAGE_CREATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_EXTERNAL_MEMORY_IMAGE_CREATE_INFO_KHR
Provided by VK_KHR_external_memory

VK_STRUCTURE_TYPE_EXTERNAL_MEMORY_IMAGE_CREATE_INFO_NV
Provided by VK_NV_external_memory

VK_STRUCTURE_TYPE_EXTERNAL_SEMAPHORE_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_EXTERNAL_SEMAPHORE_PROPERTIES_KHR
Provided by VK_KHR_external_semaphore_capabilities

VK_STRUCTURE_TYPE_FENCE_CREATE_INFO

VK_STRUCTURE_TYPE_FENCE_GET_FD_INFO_KHR
Provided by VK_KHR_external_fence_fd

VK_STRUCTURE_TYPE_FENCE_GET_WIN32_HANDLE_INFO_KHR
Provided by VK_KHR_external_fence_win32

VK_STRUCTURE_TYPE_FILTER_CUBIC_IMAGE_VIEW_IMAGE_FORMAT_PROPERTIES_EXT
Provided by VK_EXT_filter_cubic

VK_STRUCTURE_TYPE_FORMAT_PROPERTIES_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_FORMAT_PROPERTIES_2_KHR
Provided by VK_KHR_get_physical_device_properties2

VK_STRUCTURE_TYPE_FRAGMENT_SHADING_RATE_ATTACHMENT_INFO_KHR
Provided by VK_KHR_fragment_shading_rate

VK_STRUCTURE_TYPE_FRAMEBUFFER_ATTACHMENTS_CREATE_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_FRAMEBUFFER_ATTACHMENTS_CREATE_INFO_KHR
Provided by VK_KHR_imageless_framebuffer

VK_STRUCTURE_TYPE_FRAMEBUFFER_ATTACHMENT_IMAGE_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_FRAMEBUFFER_ATTACHMENT_IMAGE_INFO_KHR
Provided by VK_KHR_imageless_framebuffer

VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO

VK_STRUCTURE_TYPE_FRAMEBUFFER_MIXED_SAMPLES_COMBINATION_NV
Provided by VK_NV_coverage_reduction_mode

VK_STRUCTURE_TYPE_GENERATED_COMMANDS_INFO_NV
Provided by VK_NV_device_generated_commands

VK_STRUCTURE_TYPE_GENERATED_COMMANDS_MEMORY_REQUIREMENTS_INFO_NV
Provided by VK_NV_device_generated_commands

VK_STRUCTURE_TYPE_GEOMETRY_AABB_NV
Provided by VK_NV_ray_tracing

VK_STRUCTURE_TYPE_GEOMETRY_NV
Provided by VK_NV_ray_tracing

VK_STRUCTURE_TYPE_GEOMETRY_TRIANGLES_NV
Provided by VK_NV_ray_tracing

VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO

VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_SHADER_GROUPS_CREATE_INFO_NV
Provided by VK_NV_device_generated_commands

VK_STRUCTURE_TYPE_GRAPHICS_SHADER_GROUP_CREATE_INFO_NV
Provided by VK_NV_device_generated_commands

VK_STRUCTURE_TYPE_HDR_METADATA_EXT
Provided by VK_EXT_hdr_metadata

VK_STRUCTURE_TYPE_HEADLESS_SURFACE_CREATE_INFO_EXT
Provided by VK_EXT_headless_surface

VK_STRUCTURE_TYPE_IMAGEPIPE_SURFACE_CREATE_INFO_FUCHSIA
Provided by VK_FUCHSIA_imagepipe_surface

VK_STRUCTURE_TYPE_IMAGE_BLIT_2_KHR
Provided by VK_KHR_copy_commands2

VK_STRUCTURE_TYPE_IMAGE_COPY_2_KHR
Provided by VK_KHR_copy_commands2

VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO

VK_STRUCTURE_TYPE_IMAGE_DRM_FORMAT_MODIFIER_EXPLICIT_CREATE_INFO_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_STRUCTURE_TYPE_IMAGE_DRM_FORMAT_MODIFIER_LIST_CREATE_INFO_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_STRUCTURE_TYPE_IMAGE_DRM_FORMAT_MODIFIER_PROPERTIES_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_STRUCTURE_TYPE_IMAGE_FORMAT_LIST_CREATE_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_IMAGE_FORMAT_LIST_CREATE_INFO_KHR
Provided by VK_KHR_image_format_list

VK_STRUCTURE_TYPE_IMAGE_FORMAT_PROPERTIES_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_IMAGE_FORMAT_PROPERTIES_2_KHR
Provided by VK_KHR_get_physical_device_properties2

VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER

VK_STRUCTURE_TYPE_IMAGE_MEMORY_REQUIREMENTS_INFO_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_IMAGE_MEMORY_REQUIREMENTS_INFO_2_KHR
Provided by VK_KHR_get_memory_requirements2

VK_STRUCTURE_TYPE_IMAGE_PLANE_MEMORY_REQUIREMENTS_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_IMAGE_PLANE_MEMORY_REQUIREMENTS_INFO_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_STRUCTURE_TYPE_IMAGE_RESOLVE_2_KHR
Provided by VK_KHR_copy_commands2

VK_STRUCTURE_TYPE_IMAGE_SPARSE_MEMORY_REQUIREMENTS_INFO_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_IMAGE_SPARSE_MEMORY_REQUIREMENTS_INFO_2_KHR
Provided by VK_KHR_get_memory_requirements2

VK_STRUCTURE_TYPE_IMAGE_STENCIL_USAGE_CREATE_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_IMAGE_STENCIL_USAGE_CREATE_INFO_EXT
Provided by VK_EXT_separate_stencil_usage

VK_STRUCTURE_TYPE_IMAGE_SWAPCHAIN_CREATE_INFO_KHR
Provided by VK_KHR_swapchain with VK_VERSION_1_1, VK_KHR_device_group with VK_KHR_swapchain

VK_STRUCTURE_TYPE_IMAGE_VIEW_ADDRESS_PROPERTIES_NVX
Provided by VK_NVX_image_view_handle

VK_STRUCTURE_TYPE_IMAGE_VIEW_ASTC_DECODE_MODE_EXT
Provided by VK_EXT_astc_decode_mode

VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO

VK_STRUCTURE_TYPE_IMAGE_VIEW_HANDLE_INFO_NVX
Provided by VK_NVX_image_view_handle

VK_STRUCTURE_TYPE_IMAGE_VIEW_USAGE_CREATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_IMAGE_VIEW_USAGE_CREATE_INFO_KHR
Provided by VK_KHR_maintenance2

VK_STRUCTURE_TYPE_IMPORT_ANDROID_HARDWARE_BUFFER_INFO_ANDROID
Provided by VK_ANDROID_external_memory_android_hardware_buffer

VK_STRUCTURE_TYPE_IMPORT_FENCE_FD_INFO_KHR
Provided by VK_KHR_external_fence_fd

VK_STRUCTURE_TYPE_IMPORT_FENCE_WIN32_HANDLE_INFO_KHR
Provided by VK_KHR_external_fence_win32

VK_STRUCTURE_TYPE_IMPORT_MEMORY_FD_INFO_KHR
Provided by VK_KHR_external_memory_fd

VK_STRUCTURE_TYPE_IMPORT_MEMORY_HOST_POINTER_INFO_EXT
Provided by VK_EXT_external_memory_host

VK_STRUCTURE_TYPE_IMPORT_MEMORY_WIN32_HANDLE_INFO_KHR
Provided by VK_KHR_external_memory_win32

VK_STRUCTURE_TYPE_IMPORT_MEMORY_WIN32_HANDLE_INFO_NV
Provided by VK_NV_external_memory_win32

VK_STRUCTURE_TYPE_IMPORT_SEMAPHORE_FD_INFO_KHR
Provided by VK_KHR_external_semaphore_fd

VK_STRUCTURE_TYPE_IMPORT_SEMAPHORE_WIN32_HANDLE_INFO_KHR
Provided by VK_KHR_external_semaphore_win32

VK_STRUCTURE_TYPE_INDIRECT_COMMANDS_LAYOUT_CREATE_INFO_NV
Provided by VK_NV_device_generated_commands

VK_STRUCTURE_TYPE_INDIRECT_COMMANDS_LAYOUT_TOKEN_NV
Provided by VK_NV_device_generated_commands

VK_STRUCTURE_TYPE_INITIALIZE_PERFORMANCE_API_INFO_INTEL
Provided by VK_INTEL_performance_query

VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO

VK_STRUCTURE_TYPE_IOS_SURFACE_CREATE_INFO_MVK
Provided by VK_MVK_ios_surface

VK_STRUCTURE_TYPE_LOADER_DEVICE_CREATE_INFO
Reserved for internal use by the loader, layers, and ICDs

VK_STRUCTURE_TYPE_LOADER_INSTANCE_CREATE_INFO
Reserved for internal use by the loader, layers, and ICDs

VK_STRUCTURE_TYPE_MACOS_SURFACE_CREATE_INFO_MVK
Provided by VK_MVK_macos_surface

VK_STRUCTURE_TYPE_MAPPED_MEMORY_RANGE

VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_FLAGS_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_FLAGS_INFO_KHR
Provided by VK_KHR_device_group

VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO

VK_STRUCTURE_TYPE_MEMORY_BARRIER

VK_STRUCTURE_TYPE_MEMORY_DEDICATED_ALLOCATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_MEMORY_DEDICATED_ALLOCATE_INFO_KHR
Provided by VK_KHR_dedicated_allocation

VK_STRUCTURE_TYPE_MEMORY_DEDICATED_REQUIREMENTS
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_MEMORY_DEDICATED_REQUIREMENTS_KHR
Provided by VK_KHR_dedicated_allocation

VK_STRUCTURE_TYPE_MEMORY_FD_PROPERTIES_KHR
Provided by VK_KHR_external_memory_fd

VK_STRUCTURE_TYPE_MEMORY_GET_ANDROID_HARDWARE_BUFFER_INFO_ANDROID
Provided by VK_ANDROID_external_memory_android_hardware_buffer

VK_STRUCTURE_TYPE_MEMORY_GET_FD_INFO_KHR
Provided by VK_KHR_external_memory_fd

VK_STRUCTURE_TYPE_MEMORY_GET_WIN32_HANDLE_INFO_KHR
Provided by VK_KHR_external_memory_win32

VK_STRUCTURE_TYPE_MEMORY_HOST_POINTER_PROPERTIES_EXT
Provided by VK_EXT_external_memory_host

VK_STRUCTURE_TYPE_MEMORY_OPAQUE_CAPTURE_ADDRESS_ALLOCATE_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_MEMORY_OPAQUE_CAPTURE_ADDRESS_ALLOCATE_INFO_KHR
Provided by VK_KHR_buffer_device_address

VK_STRUCTURE_TYPE_MEMORY_PRIORITY_ALLOCATE_INFO_EXT
Provided by VK_EXT_memory_priority

VK_STRUCTURE_TYPE_MEMORY_REQUIREMENTS_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_MEMORY_REQUIREMENTS_2_KHR
Provided by VK_KHR_get_memory_requirements2

VK_STRUCTURE_TYPE_MEMORY_WIN32_HANDLE_PROPERTIES_KHR
Provided by VK_KHR_external_memory_win32

VK_STRUCTURE_TYPE_METAL_SURFACE_CREATE_INFO_EXT
Provided by VK_EXT_metal_surface

VK_STRUCTURE_TYPE_MULTISAMPLE_PROPERTIES_EXT
Provided by VK_EXT_sample_locations

VK_STRUCTURE_TYPE_MUTABLE_DESCRIPTOR_TYPE_CREATE_INFO_VALVE
Provided by VK_VALVE_mutable_descriptor_type

VK_STRUCTURE_TYPE_PERFORMANCE_CONFIGURATION_ACQUIRE_INFO_INTEL
Provided by VK_INTEL_performance_query

VK_STRUCTURE_TYPE_PERFORMANCE_COUNTER_DESCRIPTION_KHR
Provided by VK_KHR_performance_query

VK_STRUCTURE_TYPE_PERFORMANCE_COUNTER_KHR
Provided by VK_KHR_performance_query

VK_STRUCTURE_TYPE_PERFORMANCE_MARKER_INFO_INTEL
Provided by VK_INTEL_performance_query

VK_STRUCTURE_TYPE_PERFORMANCE_OVERRIDE_INFO_INTEL
Provided by VK_INTEL_performance_query

VK_STRUCTURE_TYPE_PERFORMANCE_QUERY_SUBMIT_INFO_KHR
Provided by VK_KHR_performance_query

VK_STRUCTURE_TYPE_PERFORMANCE_STREAM_MARKER_INFO_INTEL
Provided by VK_INTEL_performance_query

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_8BIT_STORAGE_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_8BIT_STORAGE_FEATURES_KHR
Provided by VK_KHR_8bit_storage

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_16BIT_STORAGE_FEATURES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_16BIT_STORAGE_FEATURES_KHR
Provided by VK_KHR_16bit_storage

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_4444_FORMATS_FEATURES_EXT
Provided by VK_EXT_4444_formats

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_ACCELERATION_STRUCTURE_FEATURES_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_ACCELERATION_STRUCTURE_PROPERTIES_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_ASTC_DECODE_FEATURES_EXT
Provided by VK_EXT_astc_decode_mode

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_BLEND_OPERATION_ADVANCED_FEATURES_EXT
Provided by VK_EXT_blend_operation_advanced

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_BLEND_OPERATION_ADVANCED_PROPERTIES_EXT
Provided by VK_EXT_blend_operation_advanced

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_BUFFER_ADDRESS_FEATURES_EXT

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_BUFFER_DEVICE_ADDRESS_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_BUFFER_DEVICE_ADDRESS_FEATURES_EXT
Provided by VK_EXT_buffer_device_address

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_BUFFER_DEVICE_ADDRESS_FEATURES_KHR
Provided by VK_KHR_buffer_device_address

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_COHERENT_MEMORY_FEATURES_AMD
Provided by VK_AMD_device_coherent_memory

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_COMPUTE_SHADER_DERIVATIVES_FEATURES_NV
Provided by VK_NV_compute_shader_derivatives

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_CONDITIONAL_RENDERING_FEATURES_EXT
Provided by VK_EXT_conditional_rendering

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_CONSERVATIVE_RASTERIZATION_PROPERTIES_EXT
Provided by VK_EXT_conservative_rasterization

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_COOPERATIVE_MATRIX_FEATURES_NV
Provided by VK_NV_cooperative_matrix

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_COOPERATIVE_MATRIX_PROPERTIES_NV
Provided by VK_NV_cooperative_matrix

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_CORNER_SAMPLED_IMAGE_FEATURES_NV
Provided by VK_NV_corner_sampled_image

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_COVERAGE_REDUCTION_MODE_FEATURES_NV
Provided by VK_NV_coverage_reduction_mode

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_CUSTOM_BORDER_COLOR_FEATURES_EXT
Provided by VK_EXT_custom_border_color

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_CUSTOM_BORDER_COLOR_PROPERTIES_EXT
Provided by VK_EXT_custom_border_color

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DEDICATED_ALLOCATION_IMAGE_ALIASING_FEATURES_NV
Provided by VK_NV_dedicated_allocation_image_aliasing

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_CLIP_ENABLE_FEATURES_EXT
Provided by VK_EXT_depth_clip_enable

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_STENCIL_RESOLVE_PROPERTIES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_STENCIL_RESOLVE_PROPERTIES_KHR
Provided by VK_KHR_depth_stencil_resolve

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DESCRIPTOR_INDEXING_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DESCRIPTOR_INDEXING_FEATURES_EXT
Provided by VK_EXT_descriptor_indexing

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DESCRIPTOR_INDEXING_PROPERTIES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DESCRIPTOR_INDEXING_PROPERTIES_EXT
Provided by VK_EXT_descriptor_indexing

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DEVICE_GENERATED_COMMANDS_FEATURES_NV
Provided by VK_NV_device_generated_commands

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DEVICE_GENERATED_COMMANDS_PROPERTIES_NV
Provided by VK_NV_device_generated_commands

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DEVICE_MEMORY_REPORT_FEATURES_EXT
Provided by VK_EXT_device_memory_report

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DIAGNOSTICS_CONFIG_FEATURES_NV
Provided by VK_NV_device_diagnostics_config

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DISCARD_RECTANGLE_PROPERTIES_EXT
Provided by VK_EXT_discard_rectangles

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DRIVER_PROPERTIES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DRIVER_PROPERTIES_KHR
Provided by VK_KHR_driver_properties

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXCLUSIVE_SCISSOR_FEATURES_NV
Provided by VK_NV_scissor_exclusive

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTENDED_DYNAMIC_STATE_FEATURES_EXT
Provided by VK_EXT_extended_dynamic_state

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_BUFFER_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_BUFFER_INFO_KHR
Provided by VK_KHR_external_memory_capabilities

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_FENCE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_FENCE_INFO_KHR
Provided by VK_KHR_external_fence_capabilities

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_IMAGE_FORMAT_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_IMAGE_FORMAT_INFO_KHR
Provided by VK_KHR_external_memory_capabilities

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_MEMORY_HOST_PROPERTIES_EXT
Provided by VK_EXT_external_memory_host

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_SEMAPHORE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_SEMAPHORE_INFO_KHR
Provided by VK_KHR_external_semaphore_capabilities

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FEATURES_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FEATURES_2_KHR
Provided by VK_KHR_get_physical_device_properties2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FLOAT16_INT8_FEATURES_KHR
Provided by VK_KHR_shader_float16_int8

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FLOAT_CONTROLS_PROPERTIES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FLOAT_CONTROLS_PROPERTIES_KHR
Provided by VK_KHR_shader_float_controls

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_DENSITY_MAP_2_FEATURES_EXT
Provided by VK_EXT_fragment_density_map2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_DENSITY_MAP_2_PROPERTIES_EXT
Provided by VK_EXT_fragment_density_map2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_DENSITY_MAP_FEATURES_EXT
Provided by VK_EXT_fragment_density_map

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_DENSITY_MAP_PROPERTIES_EXT
Provided by VK_EXT_fragment_density_map

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_SHADER_BARYCENTRIC_FEATURES_NV
Provided by VK_NV_fragment_shader_barycentric

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_SHADER_INTERLOCK_FEATURES_EXT
Provided by VK_EXT_fragment_shader_interlock

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_SHADING_RATE_ENUMS_FEATURES_NV
Provided by VK_NV_fragment_shading_rate_enums

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_SHADING_RATE_ENUMS_PROPERTIES_NV
Provided by VK_NV_fragment_shading_rate_enums

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_SHADING_RATE_FEATURES_KHR
Provided by VK_KHR_fragment_shading_rate

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_SHADING_RATE_KHR
Provided by VK_KHR_fragment_shading_rate

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_SHADING_RATE_PROPERTIES_KHR
Provided by VK_KHR_fragment_shading_rate

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_GROUP_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_GROUP_PROPERTIES_KHR
Provided by VK_KHR_device_group_creation

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_HOST_QUERY_RESET_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_HOST_QUERY_RESET_FEATURES_EXT
Provided by VK_EXT_host_query_reset

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_ID_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_ID_PROPERTIES_KHR
Provided by VK_KHR_external_memory_capabilities; VK_KHR_external_semaphore_capabilities; VK_KHR_external_fence_capabilities

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGELESS_FRAMEBUFFER_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGELESS_FRAMEBUFFER_FEATURES_KHR
Provided by VK_KHR_imageless_framebuffer

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_DRM_FORMAT_MODIFIER_INFO_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_FORMAT_INFO_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_FORMAT_INFO_2_KHR
Provided by VK_KHR_get_physical_device_properties2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_ROBUSTNESS_FEATURES_EXT
Provided by VK_EXT_image_robustness

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_VIEW_IMAGE_FORMAT_INFO_EXT
Provided by VK_EXT_filter_cubic

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_INDEX_TYPE_UINT8_FEATURES_EXT
Provided by VK_EXT_index_type_uint8

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_INLINE_UNIFORM_BLOCK_FEATURES_EXT
Provided by VK_EXT_inline_uniform_block

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_INLINE_UNIFORM_BLOCK_PROPERTIES_EXT
Provided by VK_EXT_inline_uniform_block

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_LINE_RASTERIZATION_FEATURES_EXT
Provided by VK_EXT_line_rasterization

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_LINE_RASTERIZATION_PROPERTIES_EXT
Provided by VK_EXT_line_rasterization

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MAINTENANCE_3_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MAINTENANCE_3_PROPERTIES_KHR
Provided by VK_KHR_maintenance3

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MEMORY_BUDGET_PROPERTIES_EXT
Provided by VK_EXT_memory_budget

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MEMORY_PRIORITY_FEATURES_EXT
Provided by VK_EXT_memory_priority

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MEMORY_PROPERTIES_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MEMORY_PROPERTIES_2_KHR
Provided by VK_KHR_get_physical_device_properties2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MESH_SHADER_FEATURES_NV
Provided by VK_NV_mesh_shader

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MESH_SHADER_PROPERTIES_NV
Provided by VK_NV_mesh_shader

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MULTIVIEW_FEATURES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MULTIVIEW_FEATURES_KHR
Provided by VK_KHR_multiview

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MULTIVIEW_PER_VIEW_ATTRIBUTES_PROPERTIES_NVX
Provided by VK_NVX_multiview_per_view_attributes

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MULTIVIEW_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MULTIVIEW_PROPERTIES_KHR
Provided by VK_KHR_multiview

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MUTABLE_DESCRIPTOR_TYPE_FEATURES_VALVE
Provided by VK_VALVE_mutable_descriptor_type

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PCI_BUS_INFO_PROPERTIES_EXT
Provided by VK_EXT_pci_bus_info

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PERFORMANCE_QUERY_FEATURES_KHR
Provided by VK_KHR_performance_query

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PERFORMANCE_QUERY_PROPERTIES_KHR
Provided by VK_KHR_performance_query

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PIPELINE_CREATION_CACHE_CONTROL_FEATURES_EXT
Provided by VK_EXT_pipeline_creation_cache_control

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PIPELINE_EXECUTABLE_PROPERTIES_FEATURES_KHR
Provided by VK_KHR_pipeline_executable_properties

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_POINT_CLIPPING_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_POINT_CLIPPING_PROPERTIES_KHR
Provided by VK_KHR_maintenance2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PORTABILITY_SUBSET_FEATURES_KHR
Provided by VK_KHR_portability_subset

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PORTABILITY_SUBSET_PROPERTIES_KHR
Provided by VK_KHR_portability_subset

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PRIVATE_DATA_FEATURES_EXT
Provided by VK_EXT_private_data

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROPERTIES_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROPERTIES_2_KHR
Provided by VK_KHR_get_physical_device_properties2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROTECTED_MEMORY_FEATURES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROTECTED_MEMORY_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PUSH_DESCRIPTOR_PROPERTIES_KHR
Provided by VK_KHR_push_descriptor

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_RAY_QUERY_FEATURES_KHR
Provided by VK_KHR_ray_query

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_RAY_TRACING_PIPELINE_FEATURES_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_RAY_TRACING_PIPELINE_PROPERTIES_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_RAY_TRACING_PROPERTIES_NV
Provided by VK_NV_ray_tracing

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_REPRESENTATIVE_FRAGMENT_TEST_FEATURES_NV
Provided by VK_NV_representative_fragment_test

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_ROBUSTNESS_2_FEATURES_EXT
Provided by VK_EXT_robustness2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_ROBUSTNESS_2_PROPERTIES_EXT
Provided by VK_EXT_robustness2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_FILTER_MINMAX_PROPERTIES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_FILTER_MINMAX_PROPERTIES_EXT
Provided by VK_EXT_sampler_filter_minmax

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_YCBCR_CONVERSION_FEATURES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_YCBCR_CONVERSION_FEATURES_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLE_LOCATIONS_PROPERTIES_EXT
Provided by VK_EXT_sample_locations

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SCALAR_BLOCK_LAYOUT_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SCALAR_BLOCK_LAYOUT_FEATURES_EXT
Provided by VK_EXT_scalar_block_layout

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SEPARATE_DEPTH_STENCIL_LAYOUTS_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SEPARATE_DEPTH_STENCIL_LAYOUTS_FEATURES_KHR
Provided by VK_KHR_separate_depth_stencil_layouts

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_ATOMIC_FLOAT_FEATURES_EXT
Provided by VK_EXT_shader_atomic_float

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_ATOMIC_INT64_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_ATOMIC_INT64_FEATURES_KHR
Provided by VK_KHR_shader_atomic_int64

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_CLOCK_FEATURES_KHR
Provided by VK_KHR_shader_clock

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_CORE_PROPERTIES_2_AMD
Provided by VK_AMD_shader_core_properties2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_CORE_PROPERTIES_AMD
Provided by VK_AMD_shader_core_properties

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_DEMOTE_TO_HELPER_INVOCATION_FEATURES_EXT
Provided by VK_EXT_shader_demote_to_helper_invocation

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_DRAW_PARAMETERS_FEATURES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_DRAW_PARAMETER_FEATURES

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_FLOAT16_INT8_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_FLOAT16_INT8_FEATURES_KHR
Provided by VK_KHR_shader_float16_int8

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_IMAGE_ATOMIC_INT64_FEATURES_EXT
Provided by VK_EXT_shader_image_atomic_int64

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_IMAGE_FOOTPRINT_FEATURES_NV
Provided by VK_NV_shader_image_footprint

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_INTEGER_FUNCTIONS_2_FEATURES_INTEL
Provided by VK_INTEL_shader_integer_functions2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_SM_BUILTINS_FEATURES_NV
Provided by VK_NV_shader_sm_builtins

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_SM_BUILTINS_PROPERTIES_NV
Provided by VK_NV_shader_sm_builtins

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_SUBGROUP_EXTENDED_TYPES_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_SUBGROUP_EXTENDED_TYPES_FEATURES_KHR
Provided by VK_KHR_shader_subgroup_extended_types

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_TERMINATE_INVOCATION_FEATURES_KHR
Provided by VK_KHR_shader_terminate_invocation

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADING_RATE_IMAGE_FEATURES_NV
Provided by VK_NV_shading_rate_image

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADING_RATE_IMAGE_PROPERTIES_NV
Provided by VK_NV_shading_rate_image

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SPARSE_IMAGE_FORMAT_INFO_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SPARSE_IMAGE_FORMAT_INFO_2_KHR
Provided by VK_KHR_get_physical_device_properties2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SUBGROUP_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SUBGROUP_SIZE_CONTROL_FEATURES_EXT
Provided by VK_EXT_subgroup_size_control

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SUBGROUP_SIZE_CONTROL_PROPERTIES_EXT
Provided by VK_EXT_subgroup_size_control

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SURFACE_INFO_2_KHR
Provided by VK_KHR_get_surface_capabilities2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_TEXEL_BUFFER_ALIGNMENT_FEATURES_EXT
Provided by VK_EXT_texel_buffer_alignment

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_TEXEL_BUFFER_ALIGNMENT_PROPERTIES_EXT
Provided by VK_EXT_texel_buffer_alignment

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_TEXTURE_COMPRESSION_ASTC_HDR_FEATURES_EXT
Provided by VK_EXT_texture_compression_astc_hdr

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_TIMELINE_SEMAPHORE_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_TIMELINE_SEMAPHORE_FEATURES_KHR
Provided by VK_KHR_timeline_semaphore

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_TIMELINE_SEMAPHORE_PROPERTIES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_TIMELINE_SEMAPHORE_PROPERTIES_KHR
Provided by VK_KHR_timeline_semaphore

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_TOOL_PROPERTIES_EXT
Provided by VK_EXT_tooling_info

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_TRANSFORM_FEEDBACK_FEATURES_EXT
Provided by VK_EXT_transform_feedback

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_TRANSFORM_FEEDBACK_PROPERTIES_EXT
Provided by VK_EXT_transform_feedback

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_UNIFORM_BUFFER_STANDARD_LAYOUT_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_UNIFORM_BUFFER_STANDARD_LAYOUT_FEATURES_KHR
Provided by VK_KHR_uniform_buffer_standard_layout

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VARIABLE_POINTERS_FEATURES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VARIABLE_POINTERS_FEATURES_KHR
Provided by VK_KHR_variable_pointers

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VARIABLE_POINTER_FEATURES

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VARIABLE_POINTER_FEATURES_KHR

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VERTEX_ATTRIBUTE_DIVISOR_FEATURES_EXT
Provided by VK_EXT_vertex_attribute_divisor

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VERTEX_ATTRIBUTE_DIVISOR_PROPERTIES_EXT
Provided by VK_EXT_vertex_attribute_divisor

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_1_1_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_1_1_PROPERTIES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_1_2_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_1_2_PROPERTIES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_MEMORY_MODEL_FEATURES
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_MEMORY_MODEL_FEATURES_KHR
Provided by VK_KHR_vulkan_memory_model

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_WORKGROUP_MEMORY_EXPLICIT_LAYOUT_FEATURES_KHR
Provided by VK_KHR_workgroup_memory_explicit_layout

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_YCBCR_IMAGE_ARRAYS_FEATURES_EXT
Provided by VK_EXT_ycbcr_image_arrays

VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_ZERO_INITIALIZE_WORKGROUP_MEMORY_FEATURES_KHR
Provided by VK_KHR_zero_initialize_workgroup_memory

VK_STRUCTURE_TYPE_PIPELINE_CACHE_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_ADVANCED_STATE_CREATE_INFO_EXT
Provided by VK_EXT_blend_operation_advanced

VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_COMPILER_CONTROL_CREATE_INFO_AMD
Provided by VK_AMD_pipeline_compiler_control

VK_STRUCTURE_TYPE_PIPELINE_COVERAGE_MODULATION_STATE_CREATE_INFO_NV
Provided by VK_NV_framebuffer_mixed_samples

VK_STRUCTURE_TYPE_PIPELINE_COVERAGE_REDUCTION_STATE_CREATE_INFO_NV
Provided by VK_NV_coverage_reduction_mode

VK_STRUCTURE_TYPE_PIPELINE_COVERAGE_TO_COLOR_STATE_CREATE_INFO_NV
Provided by VK_NV_fragment_coverage_to_color

VK_STRUCTURE_TYPE_PIPELINE_CREATION_FEEDBACK_CREATE_INFO_EXT
Provided by VK_EXT_pipeline_creation_feedback

VK_STRUCTURE_TYPE_PIPELINE_DEPTH_STENCIL_STATE_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_DISCARD_RECTANGLE_STATE_CREATE_INFO_EXT
Provided by VK_EXT_discard_rectangles

VK_STRUCTURE_TYPE_PIPELINE_DYNAMIC_STATE_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_EXECUTABLE_INFO_KHR
Provided by VK_KHR_pipeline_executable_properties

VK_STRUCTURE_TYPE_PIPELINE_EXECUTABLE_INTERNAL_REPRESENTATION_KHR
Provided by VK_KHR_pipeline_executable_properties

VK_STRUCTURE_TYPE_PIPELINE_EXECUTABLE_PROPERTIES_KHR
Provided by VK_KHR_pipeline_executable_properties

VK_STRUCTURE_TYPE_PIPELINE_EXECUTABLE_STATISTIC_KHR
Provided by VK_KHR_pipeline_executable_properties

VK_STRUCTURE_TYPE_PIPELINE_FRAGMENT_SHADING_RATE_ENUM_STATE_CREATE_INFO_NV
Provided by VK_NV_fragment_shading_rate_enums

VK_STRUCTURE_TYPE_PIPELINE_FRAGMENT_SHADING_RATE_STATE_CREATE_INFO_KHR
Provided by VK_KHR_fragment_shading_rate

VK_STRUCTURE_TYPE_PIPELINE_INFO_KHR
Provided by VK_KHR_pipeline_executable_properties

VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_LIBRARY_CREATE_INFO_KHR
Provided by VK_KHR_pipeline_library

VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_CONSERVATIVE_STATE_CREATE_INFO_EXT
Provided by VK_EXT_conservative_rasterization

VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_DEPTH_CLIP_STATE_CREATE_INFO_EXT
Provided by VK_EXT_depth_clip_enable

VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_LINE_STATE_CREATE_INFO_EXT
Provided by VK_EXT_line_rasterization

VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_RASTERIZATION_ORDER_AMD
Provided by VK_AMD_rasterization_order

VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_STREAM_CREATE_INFO_EXT
Provided by VK_EXT_transform_feedback

VK_STRUCTURE_TYPE_PIPELINE_REPRESENTATIVE_FRAGMENT_TEST_STATE_CREATE_INFO_NV
Provided by VK_NV_representative_fragment_test

VK_STRUCTURE_TYPE_PIPELINE_SAMPLE_LOCATIONS_STATE_CREATE_INFO_EXT
Provided by VK_EXT_sample_locations

VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_REQUIRED_SUBGROUP_SIZE_CREATE_INFO_EXT
Provided by VK_EXT_subgroup_size_control

VK_STRUCTURE_TYPE_PIPELINE_TESSELLATION_DOMAIN_ORIGIN_STATE_CREATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_PIPELINE_TESSELLATION_DOMAIN_ORIGIN_STATE_CREATE_INFO_KHR
Provided by VK_KHR_maintenance2

VK_STRUCTURE_TYPE_PIPELINE_TESSELLATION_STATE_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_DIVISOR_STATE_CREATE_INFO_EXT
Provided by VK_EXT_vertex_attribute_divisor

VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_COARSE_SAMPLE_ORDER_STATE_CREATE_INFO_NV
Provided by VK_NV_shading_rate_image

VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_EXCLUSIVE_SCISSOR_STATE_CREATE_INFO_NV
Provided by VK_NV_scissor_exclusive

VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_SHADING_RATE_IMAGE_STATE_CREATE_INFO_NV
Provided by VK_NV_shading_rate_image

VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO

VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_SWIZZLE_STATE_CREATE_INFO_NV
Provided by VK_NV_viewport_swizzle

VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_W_SCALING_STATE_CREATE_INFO_NV
Provided by VK_NV_clip_space_w_scaling

VK_STRUCTURE_TYPE_PRESENT_FRAME_TOKEN_GGP
Provided by VK_GGP_frame_token

VK_STRUCTURE_TYPE_PRESENT_INFO_KHR
Provided by VK_KHR_swapchain

VK_STRUCTURE_TYPE_PRESENT_REGIONS_KHR
Provided by VK_KHR_incremental_present

VK_STRUCTURE_TYPE_PRESENT_TIMES_INFO_GOOGLE
Provided by VK_GOOGLE_display_timing

VK_STRUCTURE_TYPE_PRIVATE_DATA_SLOT_CREATE_INFO_EXT
Provided by VK_EXT_private_data

VK_STRUCTURE_TYPE_PROTECTED_SUBMIT_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_QUERY_POOL_CREATE_INFO

VK_STRUCTURE_TYPE_QUERY_POOL_CREATE_INFO_INTEL

VK_STRUCTURE_TYPE_QUERY_POOL_PERFORMANCE_CREATE_INFO_KHR
Provided by VK_KHR_performance_query

VK_STRUCTURE_TYPE_QUERY_POOL_PERFORMANCE_QUERY_CREATE_INFO_INTEL
Provided by VK_INTEL_performance_query

VK_STRUCTURE_TYPE_QUEUE_FAMILY_CHECKPOINT_PROPERTIES_NV
Provided by VK_NV_device_diagnostic_checkpoints

VK_STRUCTURE_TYPE_QUEUE_FAMILY_PROPERTIES_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_QUEUE_FAMILY_PROPERTIES_2_KHR
Provided by VK_KHR_get_physical_device_properties2

VK_STRUCTURE_TYPE_RAY_TRACING_PIPELINE_CREATE_INFO_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_STRUCTURE_TYPE_RAY_TRACING_PIPELINE_CREATE_INFO_NV
Provided by VK_NV_ray_tracing

VK_STRUCTURE_TYPE_RAY_TRACING_PIPELINE_INTERFACE_CREATE_INFO_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_STRUCTURE_TYPE_RAY_TRACING_SHADER_GROUP_CREATE_INFO_KHR
Provided by VK_KHR_ray_tracing_pipeline

VK_STRUCTURE_TYPE_RAY_TRACING_SHADER_GROUP_CREATE_INFO_NV
Provided by VK_NV_ray_tracing

VK_STRUCTURE_TYPE_RENDER_PASS_ATTACHMENT_BEGIN_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_RENDER_PASS_ATTACHMENT_BEGIN_INFO_KHR
Provided by VK_KHR_imageless_framebuffer

VK_STRUCTURE_TYPE_RENDER_PASS_BEGIN_INFO

VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO

VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO_2
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO_2_KHR
Provided by VK_KHR_create_renderpass2

VK_STRUCTURE_TYPE_RENDER_PASS_FRAGMENT_DENSITY_MAP_CREATE_INFO_EXT
Provided by VK_EXT_fragment_density_map

VK_STRUCTURE_TYPE_RENDER_PASS_INPUT_ATTACHMENT_ASPECT_CREATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_RENDER_PASS_INPUT_ATTACHMENT_ASPECT_CREATE_INFO_KHR
Provided by VK_KHR_maintenance2

VK_STRUCTURE_TYPE_RENDER_PASS_MULTIVIEW_CREATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_RENDER_PASS_MULTIVIEW_CREATE_INFO_KHR
Provided by VK_KHR_multiview

VK_STRUCTURE_TYPE_RENDER_PASS_SAMPLE_LOCATIONS_BEGIN_INFO_EXT
Provided by VK_EXT_sample_locations

VK_STRUCTURE_TYPE_RENDER_PASS_TRANSFORM_BEGIN_INFO_QCOM
Provided by VK_QCOM_render_pass_transform

VK_STRUCTURE_TYPE_RESOLVE_IMAGE_INFO_2_KHR
Provided by VK_KHR_copy_commands2

VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO

VK_STRUCTURE_TYPE_SAMPLER_CUSTOM_BORDER_COLOR_CREATE_INFO_EXT
Provided by VK_EXT_custom_border_color

VK_STRUCTURE_TYPE_SAMPLER_REDUCTION_MODE_CREATE_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_SAMPLER_REDUCTION_MODE_CREATE_INFO_EXT
Provided by VK_EXT_sampler_filter_minmax

VK_STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_CREATE_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_CREATE_INFO_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_IMAGE_FORMAT_PROPERTIES
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_IMAGE_FORMAT_PROPERTIES_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_INFO
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_INFO_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_STRUCTURE_TYPE_SAMPLE_LOCATIONS_INFO_EXT
Provided by VK_EXT_sample_locations

VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO

VK_STRUCTURE_TYPE_SEMAPHORE_GET_FD_INFO_KHR
Provided by VK_KHR_external_semaphore_fd

VK_STRUCTURE_TYPE_SEMAPHORE_GET_WIN32_HANDLE_INFO_KHR
Provided by VK_KHR_external_semaphore_win32

VK_STRUCTURE_TYPE_SEMAPHORE_SIGNAL_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_SEMAPHORE_SIGNAL_INFO_KHR
Provided by VK_KHR_timeline_semaphore

VK_STRUCTURE_TYPE_SEMAPHORE_TYPE_CREATE_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_SEMAPHORE_TYPE_CREATE_INFO_KHR
Provided by VK_KHR_timeline_semaphore

VK_STRUCTURE_TYPE_SEMAPHORE_WAIT_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_SEMAPHORE_WAIT_INFO_KHR
Provided by VK_KHR_timeline_semaphore

VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO

VK_STRUCTURE_TYPE_SHADER_MODULE_VALIDATION_CACHE_CREATE_INFO_EXT
Provided by VK_EXT_validation_cache

VK_STRUCTURE_TYPE_SHARED_PRESENT_SURFACE_CAPABILITIES_KHR
Provided by VK_KHR_shared_presentable_image

VK_STRUCTURE_TYPE_SPARSE_IMAGE_FORMAT_PROPERTIES_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_SPARSE_IMAGE_FORMAT_PROPERTIES_2_KHR
Provided by VK_KHR_get_physical_device_properties2

VK_STRUCTURE_TYPE_SPARSE_IMAGE_MEMORY_REQUIREMENTS_2
Provided by VK_VERSION_1_1

VK_STRUCTURE_TYPE_SPARSE_IMAGE_MEMORY_REQUIREMENTS_2_KHR
Provided by VK_KHR_get_memory_requirements2

VK_STRUCTURE_TYPE_STREAM_DESCRIPTOR_SURFACE_CREATE_INFO_GGP
Provided by VK_GGP_stream_descriptor_surface

VK_STRUCTURE_TYPE_SUBMIT_INFO

VK_STRUCTURE_TYPE_SUBPASS_BEGIN_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_SUBPASS_BEGIN_INFO_KHR
Provided by VK_KHR_create_renderpass2

VK_STRUCTURE_TYPE_SUBPASS_DEPENDENCY_2
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_SUBPASS_DEPENDENCY_2_KHR
Provided by VK_KHR_create_renderpass2

VK_STRUCTURE_TYPE_SUBPASS_DESCRIPTION_2
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_SUBPASS_DESCRIPTION_2_KHR
Provided by VK_KHR_create_renderpass2

VK_STRUCTURE_TYPE_SUBPASS_DESCRIPTION_DEPTH_STENCIL_RESOLVE
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_SUBPASS_DESCRIPTION_DEPTH_STENCIL_RESOLVE_KHR
Provided by VK_KHR_depth_stencil_resolve

VK_STRUCTURE_TYPE_SUBPASS_END_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_SUBPASS_END_INFO_KHR
Provided by VK_KHR_create_renderpass2

VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES2_EXT

VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES_2_EXT
Provided by VK_EXT_display_surface_counter

VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES_2_KHR
Provided by VK_KHR_get_surface_capabilities2

VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES_FULL_SCREEN_EXCLUSIVE_EXT
Provided by VK_EXT_full_screen_exclusive

VK_STRUCTURE_TYPE_SURFACE_FORMAT_2_KHR
Provided by VK_KHR_get_surface_capabilities2

VK_STRUCTURE_TYPE_SURFACE_FULL_SCREEN_EXCLUSIVE_INFO_EXT
Provided by VK_EXT_full_screen_exclusive

VK_STRUCTURE_TYPE_SURFACE_FULL_SCREEN_EXCLUSIVE_WIN32_INFO_EXT
Provided by VK_EXT_full_screen_exclusive with VK_KHR_win32_surface

VK_STRUCTURE_TYPE_SURFACE_PROTECTED_CAPABILITIES_KHR
Provided by VK_KHR_surface_protected_capabilities

VK_STRUCTURE_TYPE_SWAPCHAIN_COUNTER_CREATE_INFO_EXT
Provided by VK_EXT_display_control

VK_STRUCTURE_TYPE_SWAPCHAIN_CREATE_INFO_KHR
Provided by VK_KHR_swapchain

VK_STRUCTURE_TYPE_SWAPCHAIN_DISPLAY_NATIVE_HDR_CREATE_INFO_AMD
Provided by VK_AMD_display_native_hdr

VK_STRUCTURE_TYPE_TEXTURE_LOD_GATHER_FORMAT_PROPERTIES_AMD
Provided by VK_AMD_texture_gather_bias_lod

VK_STRUCTURE_TYPE_TIMELINE_SEMAPHORE_SUBMIT_INFO
Provided by VK_VERSION_1_2

VK_STRUCTURE_TYPE_TIMELINE_SEMAPHORE_SUBMIT_INFO_KHR
Provided by VK_KHR_timeline_semaphore

VK_STRUCTURE_TYPE_VALIDATION_CACHE_CREATE_INFO_EXT
Provided by VK_EXT_validation_cache

VK_STRUCTURE_TYPE_VALIDATION_FEATURES_EXT
Provided by VK_EXT_validation_features

VK_STRUCTURE_TYPE_VALIDATION_FLAGS_EXT
Provided by VK_EXT_validation_flags

VK_STRUCTURE_TYPE_VI_SURFACE_CREATE_INFO_NN
Provided by VK_NN_vi_surface

VK_STRUCTURE_TYPE_WAYLAND_SURFACE_CREATE_INFO_KHR
Provided by VK_KHR_wayland_surface

VK_STRUCTURE_TYPE_WIN32_KEYED_MUTEX_ACQUIRE_RELEASE_INFO_KHR
Provided by VK_KHR_win32_keyed_mutex

VK_STRUCTURE_TYPE_WIN32_KEYED_MUTEX_ACQUIRE_RELEASE_INFO_NV
Provided by VK_NV_win32_keyed_mutex

VK_STRUCTURE_TYPE_WIN32_SURFACE_CREATE_INFO_KHR
Provided by VK_KHR_win32_surface

VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET

VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET_ACCELERATION_STRUCTURE_KHR
Provided by VK_KHR_acceleration_structure

VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET_ACCELERATION_STRUCTURE_NV
Provided by VK_NV_ray_tracing

VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET_INLINE_UNIFORM_BLOCK_EXT
Provided by VK_EXT_inline_uniform_block

VK_STRUCTURE_TYPE_XCB_SURFACE_CREATE_INFO_KHR
Provided by VK_KHR_xcb_surface

VK_STRUCTURE_TYPE_XLIB_SURFACE_CREATE_INFO_KHR
Provided by VK_KHR_xlib_surface