Module vkvk::flag_bits

Vulkan types where one or more flags are bit-packed together.

These types support the standard bitwise operations for enabling and disabling specific bits:

Add a new flag to a base value: base | new

Remove a flag from a base value: base & (!removal)

Structs

VkAccessFlagBits
VkAccessFlagBits

VkImageAspectFlagBits
VkImageAspectFlagBits

Constants

VK_ACCESS_ACCELERATION_STRUCTURE_READ_BIT_KHR
Provided by VK_KHR_acceleration_structure

VK_ACCESS_ACCELERATION_STRUCTURE_READ_BIT_NV
Provided by VK_NV_ray_tracing

VK_ACCESS_ACCELERATION_STRUCTURE_WRITE_BIT_KHR
Provided by VK_KHR_acceleration_structure

VK_ACCESS_ACCELERATION_STRUCTURE_WRITE_BIT_NV
Provided by VK_NV_ray_tracing

VK_ACCESS_COLOR_ATTACHMENT_READ_BIT
Controls coherency of color attachment reads

VK_ACCESS_COLOR_ATTACHMENT_READ_NONCOHERENT_BIT_EXT
Provided by VK_EXT_blend_operation_advanced

VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT
Controls coherency of color attachment writes

VK_ACCESS_COMMAND_PREPROCESS_READ_BIT_NV
Provided by VK_NV_device_generated_commands

VK_ACCESS_COMMAND_PREPROCESS_WRITE_BIT_NV
Provided by VK_NV_device_generated_commands

VK_ACCESS_CONDITIONAL_RENDERING_READ_BIT_EXT
Provided by VK_EXT_conditional_rendering

VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_READ_BIT
Controls coherency of depth/stencil attachment reads

VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT
Controls coherency of depth/stencil attachment writes

VK_ACCESS_FRAGMENT_DENSITY_MAP_READ_BIT_EXT
Provided by VK_EXT_fragment_density_map

VK_ACCESS_FRAGMENT_SHADING_RATE_ATTACHMENT_READ_BIT_KHR
Provided by VK_KHR_fragment_shading_rate

VK_ACCESS_HOST_READ_BIT
Controls coherency of host reads

VK_ACCESS_HOST_WRITE_BIT
Controls coherency of host writes

VK_ACCESS_INDEX_READ_BIT
Controls coherency of index reads

VK_ACCESS_INDIRECT_COMMAND_READ_BIT
Controls coherency of indirect command reads

VK_ACCESS_INPUT_ATTACHMENT_READ_BIT
Controls coherency of input attachment reads

VK_ACCESS_MEMORY_READ_BIT
Controls coherency of memory reads

VK_ACCESS_MEMORY_WRITE_BIT
Controls coherency of memory writes

VK_ACCESS_NONE_KHR
Provided by VK_KHR_synchronization2

VK_ACCESS_SHADER_READ_BIT
Controls coherency of shader reads

VK_ACCESS_SHADER_WRITE_BIT
Controls coherency of shader writes

VK_ACCESS_SHADING_RATE_IMAGE_READ_BIT_NV
Provided by VK_NV_shading_rate_image

VK_ACCESS_TRANSFER_READ_BIT
Controls coherency of transfer reads

VK_ACCESS_TRANSFER_WRITE_BIT
Controls coherency of transfer writes

VK_ACCESS_TRANSFORM_FEEDBACK_COUNTER_READ_BIT_EXT
Provided by VK_EXT_transform_feedback

VK_ACCESS_TRANSFORM_FEEDBACK_COUNTER_WRITE_BIT_EXT
Provided by VK_EXT_transform_feedback

VK_ACCESS_TRANSFORM_FEEDBACK_WRITE_BIT_EXT
Provided by VK_EXT_transform_feedback

VK_ACCESS_UNIFORM_READ_BIT
Controls coherency of uniform buffer reads

VK_ACCESS_VERTEX_ATTRIBUTE_READ_BIT
Controls coherency of vertex attribute reads

VK_IMAGE_ASPECT_COLOR_BIT

VK_IMAGE_ASPECT_DEPTH_BIT

VK_IMAGE_ASPECT_MEMORY_PLANE_0_BIT_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_IMAGE_ASPECT_MEMORY_PLANE_1_BIT_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_IMAGE_ASPECT_MEMORY_PLANE_2_BIT_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_IMAGE_ASPECT_MEMORY_PLANE_3_BIT_EXT
Provided by VK_EXT_image_drm_format_modifier

VK_IMAGE_ASPECT_METADATA_BIT

VK_IMAGE_ASPECT_PLANE_0_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_ASPECT_PLANE_0_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_IMAGE_ASPECT_PLANE_1_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_ASPECT_PLANE_1_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_IMAGE_ASPECT_PLANE_2_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_ASPECT_PLANE_2_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_IMAGE_ASPECT_STENCIL_BIT

Type Definitions

VkAccessFlags

VkImageAspectFlags