Module vkvk::flag_bits

Vulkan types where one or more flags are bit-packed together.

These types support the standard bitwise operations for enabling and disabling specific bits:

Add a new flag to a base value: base | new

Remove a flag from a base value: base & (!removal)

Structs

VkAccessFlagBits
VkAccessFlagBits

VkFormatFeatureFlagBits
VkFormatFeatureFlagBits

VkImageAspectFlagBits
VkImageAspectFlagBits

VkImageCreateFlagBits
VkImageCreateFlagBits

VkImageUsageFlagBits
VkImageUsageFlagBits

VkInstanceCreateFlagBits
VkInstanceCreateFlags

VkMemoryHeapFlagBits
VkMemoryHeapFlagBits

VkMemoryPropertyFlagBits
VkMemoryPropertyFlagBits

VkQueueFlagBits
VkQueueFlagBits

VkSampleCountFlagBits
VkSampleCountFlagBits

Constants

VK_ACCESS_ACCELERATION_STRUCTURE_READ_BIT_KHR

VK_ACCESS_ACCELERATION_STRUCTURE_READ_BIT_NV

VK_ACCESS_ACCELERATION_STRUCTURE_WRITE_BIT_KHR

VK_ACCESS_ACCELERATION_STRUCTURE_WRITE_BIT_NV

VK_ACCESS_COLOR_ATTACHMENT_READ_BIT
Controls coherency of color attachment reads

VK_ACCESS_COLOR_ATTACHMENT_READ_NONCOHERENT_BIT_EXT

VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT
Controls coherency of color attachment writes

VK_ACCESS_COMMAND_PREPROCESS_READ_BIT_NV

VK_ACCESS_COMMAND_PREPROCESS_WRITE_BIT_NV

VK_ACCESS_CONDITIONAL_RENDERING_READ_BIT_EXT

VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_READ_BIT
Controls coherency of depth/stencil attachment reads

VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT
Controls coherency of depth/stencil attachment writes

VK_ACCESS_FRAGMENT_DENSITY_MAP_READ_BIT_EXT

VK_ACCESS_FRAGMENT_SHADING_RATE_ATTACHMENT_READ_BIT_KHR

VK_ACCESS_HOST_READ_BIT
Controls coherency of host reads

VK_ACCESS_HOST_WRITE_BIT
Controls coherency of host writes

VK_ACCESS_INDEX_READ_BIT
Controls coherency of index reads

VK_ACCESS_INDIRECT_COMMAND_READ_BIT
Controls coherency of indirect command reads

VK_ACCESS_INPUT_ATTACHMENT_READ_BIT
Controls coherency of input attachment reads

VK_ACCESS_MEMORY_READ_BIT
Controls coherency of memory reads

VK_ACCESS_MEMORY_WRITE_BIT
Controls coherency of memory writes

VK_ACCESS_NONE_KHR

VK_ACCESS_SHADER_READ_BIT
Controls coherency of shader reads

VK_ACCESS_SHADER_WRITE_BIT
Controls coherency of shader writes

VK_ACCESS_SHADING_RATE_IMAGE_READ_BIT_NV

VK_ACCESS_TRANSFER_READ_BIT
Controls coherency of transfer reads

VK_ACCESS_TRANSFER_WRITE_BIT
Controls coherency of transfer writes

VK_ACCESS_TRANSFORM_FEEDBACK_COUNTER_READ_BIT_EXT

VK_ACCESS_TRANSFORM_FEEDBACK_COUNTER_WRITE_BIT_EXT

VK_ACCESS_TRANSFORM_FEEDBACK_WRITE_BIT_EXT

VK_ACCESS_UNIFORM_READ_BIT
Controls coherency of uniform buffer reads

VK_ACCESS_VERTEX_ATTRIBUTE_READ_BIT
Controls coherency of vertex attribute reads

VK_FORMAT_FEATURE_ACCELERATION_STRUCTURE_VERTEX_BUFFER_BIT_KHR
Provided by VK_KHR_acceleration_structure

VK_FORMAT_FEATURE_BLIT_DST_BIT
Format can be used as the destination image of blits with vkCmdBlitImage

VK_FORMAT_FEATURE_BLIT_SRC_BIT
Format can be used as the source image of blits with vkCmdBlitImage

VK_FORMAT_FEATURE_COLOR_ATTACHMENT_BIT
Format can be used for color attachment images

VK_FORMAT_FEATURE_COLOR_ATTACHMENT_BLEND_BIT
Format supports blending in case it is used for color attachment images

VK_FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT
Format can be used for depth/stencil attachment images

VK_FORMAT_FEATURE_DISJOINT_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_DISJOINT_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_FRAGMENT_DENSITY_MAP_BIT_EXT
Provided by VK_EXT_fragment_density_map

VK_FORMAT_FEATURE_FRAGMENT_SHADING_RATE_ATTACHMENT_BIT_KHR
Provided by VK_KHR_fragment_shading_rate

VK_FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_SAMPLED_IMAGE_BIT
Format can be used for sampled images (SAMPLED_IMAGE and COMBINED_IMAGE_SAMPLER descriptor types)

VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_CUBIC_BIT_EXT
Provided by VK_EXT_filter_cubic

VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_CUBIC_BIT_IMG
Provided by VK_IMG_filter_cubic

VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT
Format can be filtered with VK_FILTER_LINEAR when being sampled

VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_MINMAX_BIT
Provided by VK_VERSION_1_2

VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_MINMAX_BIT_EXT
Provided by VK_EXT_sampler_filter_minmax

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_SEPARATE_RECONSTRUCTION_FILTER_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_SEPARATE_RECONSTRUCTION_FILTER_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_FORMAT_FEATURE_STORAGE_IMAGE_ATOMIC_BIT
Format supports atomic operations in case it is used for storage images

VK_FORMAT_FEATURE_STORAGE_IMAGE_BIT
Format can be used for storage images (STORAGE_IMAGE descriptor type)

VK_FORMAT_FEATURE_STORAGE_TEXEL_BUFFER_ATOMIC_BIT
Format supports atomic operations in case it is used for storage texel buffers

VK_FORMAT_FEATURE_STORAGE_TEXEL_BUFFER_BIT
Format can be used for storage texel buffers (IBOs)

VK_FORMAT_FEATURE_TRANSFER_DST_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_TRANSFER_DST_BIT_KHR
Provided by VK_KHR_maintenance1

VK_FORMAT_FEATURE_TRANSFER_SRC_BIT
Provided by VK_VERSION_1_1

VK_FORMAT_FEATURE_TRANSFER_SRC_BIT_KHR
Provided by VK_KHR_maintenance1

VK_FORMAT_FEATURE_UNIFORM_TEXEL_BUFFER_BIT
Format can be used for uniform texel buffers (TBOs)

VK_FORMAT_FEATURE_VERTEX_BUFFER_BIT
Format can be used for vertex buffers (VBOs)

VK_IMAGE_ASPECT_COLOR_BIT

VK_IMAGE_ASPECT_DEPTH_BIT

VK_IMAGE_ASPECT_MEMORY_PLANE_0_BIT_EXT

VK_IMAGE_ASPECT_MEMORY_PLANE_1_BIT_EXT

VK_IMAGE_ASPECT_MEMORY_PLANE_2_BIT_EXT

VK_IMAGE_ASPECT_MEMORY_PLANE_3_BIT_EXT

VK_IMAGE_ASPECT_METADATA_BIT

VK_IMAGE_ASPECT_PLANE_0_BIT

VK_IMAGE_ASPECT_PLANE_0_BIT_KHR

VK_IMAGE_ASPECT_PLANE_1_BIT

VK_IMAGE_ASPECT_PLANE_1_BIT_KHR

VK_IMAGE_ASPECT_PLANE_2_BIT

VK_IMAGE_ASPECT_PLANE_2_BIT_KHR

VK_IMAGE_ASPECT_STENCIL_BIT

VK_IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT_KHR
Provided by VK_KHR_maintenance1

VK_IMAGE_CREATE_ALIAS_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_ALIAS_BIT_KHR
Provided by VK_KHR_bind_memory2

VK_IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT_KHR
Provided by VK_KHR_maintenance2

VK_IMAGE_CREATE_CORNER_SAMPLED_BIT_NV
Provided by VK_NV_corner_sampled_image

VK_IMAGE_CREATE_CUBE_COMPATIBLE_BIT
Allows creating image views with cube type from the created image

VK_IMAGE_CREATE_DISJOINT_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_DISJOINT_BIT_KHR
Provided by VK_KHR_sampler_ycbcr_conversion

VK_IMAGE_CREATE_EXTENDED_USAGE_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_EXTENDED_USAGE_BIT_KHR
Provided by VK_KHR_maintenance2

VK_IMAGE_CREATE_MUTABLE_FORMAT_BIT
Allows image views to have different format than the base image

VK_IMAGE_CREATE_PROTECTED_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_SAMPLE_LOCATIONS_COMPATIBLE_DEPTH_BIT_EXT
Provided by VK_EXT_sample_locations

VK_IMAGE_CREATE_SPARSE_ALIASED_BIT
Image should support constant data access to physical memory ranges mapped into multiple locations of sparse images

VK_IMAGE_CREATE_SPARSE_BINDING_BIT
Image should support sparse backing

VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT
Image should support sparse backing with partial residency

VK_IMAGE_CREATE_SPLIT_INSTANCE_BIND_REGIONS_BIT
Provided by VK_VERSION_1_1

VK_IMAGE_CREATE_SPLIT_INSTANCE_BIND_REGIONS_BIT_KHR
Provided by VK_KHR_device_group with VK_KHR_bind_memory2

VK_IMAGE_CREATE_SUBSAMPLED_BIT_EXT
Provided by VK_EXT_fragment_density_map

VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT
Can be used as framebuffer color attachment

VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT
Can be used as framebuffer depth/stencil attachment

VK_IMAGE_USAGE_FRAGMENT_DENSITY_MAP_BIT_EXT
Provided by VK_EXT_fragment_density_map

VK_IMAGE_USAGE_FRAGMENT_SHADING_RATE_ATTACHMENT_BIT_KHR
Provided by VK_KHR_fragment_shading_rate

VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT
Can be used as framebuffer input attachment

VK_IMAGE_USAGE_SAMPLED_BIT
Can be sampled from (SAMPLED_IMAGE and COMBINED_IMAGE_SAMPLER descriptor types)

VK_IMAGE_USAGE_SHADING_RATE_IMAGE_BIT_NV
Provided by VK_NV_shading_rate_image

VK_IMAGE_USAGE_STORAGE_BIT
Can be used as storage image (STORAGE_IMAGE descriptor type)

VK_IMAGE_USAGE_TRANSFER_DST_BIT
Can be used as a destination of transfer operations

VK_IMAGE_USAGE_TRANSFER_SRC_BIT
Can be used as a source of transfer operations

VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT
Image data not needed outside of rendering

VK_MEMORY_HEAP_DEVICE_LOCAL_BIT
If set, heap represents device memory

VK_MEMORY_HEAP_MULTI_INSTANCE_BIT
Provided by VK_VERSION_1_1

VK_MEMORY_HEAP_MULTI_INSTANCE_BIT_KHR
Provided by VK_KHR_device_group_creation

VK_MEMORY_PROPERTY_DEVICE_COHERENT_BIT_AMD
Provided by VK_AMD_device_coherent_memory

VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT
If otherwise stated, then allocate memory on device

VK_MEMORY_PROPERTY_DEVICE_UNCACHED_BIT_AMD
Provided by VK_AMD_device_coherent_memory

VK_MEMORY_PROPERTY_HOST_CACHED_BIT
Memory will be cached by the host

VK_MEMORY_PROPERTY_HOST_COHERENT_BIT
Memory will have i/o coherency. If not set, application may need to use vkFlushMappedMemoryRanges and vkInvalidateMappedMemoryRanges to flush/invalidate host cache

VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT
Memory is mappable by host

VK_MEMORY_PROPERTY_LAZILY_ALLOCATED_BIT
Memory may be allocated by the driver when it is required

VK_MEMORY_PROPERTY_PROTECTED_BIT
Provided by VK_VERSION_1_1

VK_QUEUE_COMPUTE_BIT
Queue supports compute operations

VK_QUEUE_GRAPHICS_BIT
Queue supports graphics operations

VK_QUEUE_PROTECTED_BIT
Provided by VK_VERSION_1_1

VK_QUEUE_SPARSE_BINDING_BIT
Queue supports sparse resource memory management operations

VK_QUEUE_TRANSFER_BIT
Queue supports transfer operations

VK_SAMPLE_COUNT_1_BIT
Sample count 1 supported

VK_SAMPLE_COUNT_2_BIT
Sample count 2 supported

VK_SAMPLE_COUNT_4_BIT
Sample count 4 supported

VK_SAMPLE_COUNT_8_BIT
Sample count 8 supported

VK_SAMPLE_COUNT_16_BIT
Sample count 16 supported

VK_SAMPLE_COUNT_32_BIT
Sample count 32 supported

VK_SAMPLE_COUNT_64_BIT
Sample count 64 supported

Type Definitions

VkAccessFlags

VkFormatFeatureFlags

VkImageAspectFlags

VkImageCreateFlags

VkImageUsageFlags

VkInstanceCreateFlags

VkMemoryHeapFlags

VkMemoryPropertyFlags

VkQueueFlags

VkSampleCountFlags